% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{aa_transform}
\alias{aa_transform}
\title{account activities transform}
\usage{
aa_transform(resp)
}
\arguments{
\item{resp}{Reponse from account_activities endpoint}
}
\description{
transform account activities
}
\keyword{internal}
