% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{account_activities}
\alias{account_activities}
\title{Get Account Activity}
\usage{
account_activities(
  activity_type = NULL,
  date = NULL,
  until = NULL,
  after = NULL,
  direction = "desc",
  page_size = 50,
  page_token = NULL,
  live = as.logical(Sys.getenv("APCA-LIVE", FALSE)),
  v = 2
)
}
\arguments{
\item{activity_type}{\code{(character)} The activity type you want to view entries for. A list of valid activity types can be found in Details.}

\item{date}{\code{(character/Date)} The date in YYYY-MM-DD format for which you want to see activities.}

\item{until}{\code{(character/Date)} The response will contain only activities submitted before this date in YYYY-MM-DD format. (Cannot be used with date.)}

\item{after}{\code{(character/Date)} The response will contain only activities submitted after this date in YYYY-MM-DD format. (Cannot be used with date.)}

\item{direction}{\code{(character)} asc or desc, default is desc.}

\item{page_size}{\code{(integer/numeric)} The maximum number of entries to return in the response.}

\item{page_token}{\code{(character)} The ID of the end of your current page of results to indicate where the next page should start.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
\code{TradeActivity} \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/account-activities/#tradeactivity-entity}{TradeActivity} Object or  a \href{https://alpaca.markets/docs/api-documentation/api-v2/account-activities/#nontradeactivity-entity}{NonTradeActivity} Object. See Details.
\itemize{
\item{\code{id}}{ \code{(character)} An id for the activity. Always in "
::" format. Can be sent as \code{page_token} in requests to facilitate the paging of results.}
\item{\code{activity_type}}{ \code{(character)} Activity type to filter for. IE "FILL". Non case-sensitive.}
\item{\code{transaction_time}}{ \code{(POSIXct)} The time at which the execution occurred.}
\item{\code{type}}{ \code{(character)} \code{"fill"} or \code{"partial_fill"}}
\item{\code{price}}{ \code{(numeric)} The per-share price that the trade was executed at.}
\item{\code{qty}}{ \code{(integer)} The number of shares involved in the trade execution.}
\item{\code{side}}{ \code{(character)} \code{"buy"} or \code{"sell"}}
\item{\code{symbol}}{ \code{(character)} The symbol of the security being traded.}
\item{\code{leaves_qty}}{ \code{(integer)} For partially_filled orders, the quantity of shares that are left to be filled.}
\item{\code{order_id}}{ \code{(character)} The id for the order that filled.}
\item{\code{cum_qty}}{ \code{(integer)} The cumulative quantity of shares involved in the execution.}
}

\code{NonTradeActivity} \code{(tibble)} A \href{https://alpaca.markets/docs/api-documentation/api-v2/account-activities#nontradeactivity-entity}{NonTradeActivity Entity}:
\itemize{
\item{\code{activity_type}}{ \code{(character)} See Details}
\item{\code{id}}{ \code{(character)} An id for the activity. Always in "
::" format. Can be sent as \code{page_token} in requests to facilitate the paging of results.}
\item{\code{date}}{ \code{(POSIXct)} The date on which the activity occurred or on which the transaction associated with the activity settled.}
\item{\code{net_amount}}{ \code{(numeric)} The net amount of money (positive or negative) associated with the activity.}
\item{\code{symbol}}{ \code{(character)} The symbol of the security involved with the activity. Not present for all activity types.}
\item{\code{qty}}{ \code{(integer)} For dividend activities, the number of shares that contributed to the payment. Not present for other activity types.}
\item{\code{per_share_amount}}{ \code{(numeric)} For dividend activities, the average amount paid per share. Not present for other activity types.}
}
}
\description{
The account activities API provides access to a historical record of transaction activities that have impacted your account. Trade execution activities and non-trade activities, such as dividend payments, are both reported through this endpoint. See \href{https://alpaca.markets/docs/api-documentation/api-v2/account-activities/}{Account Activities} for details.
}
\details{
Activity Types:
\itemize{
\item{\code{"FILL"}}{ Order fills (both partial and full fills)}
\item{\code{"TRANS"}}{ Cash transactions (both CSD and CSR)}
\item{\code{"MISC"}}{ Miscellaneous or rarely used activity types (All types except those in TRANS, DIV, or   FILL)}
\item{\code{"ACATC"}}{ ACATS IN/OUT (Cash)}
\item{\code{"ACATS"}}{ ACATS IN/OUT (Securities)}
\item{\code{"CSD"}}{ Cash disbursement(+)}
\item{\code{"CSR"}}{ Cash receipt(-)}
\item{\code{"DIV"}}{ Dividends}
\item{\code{"DIVCGL"}}{ Dividend (capital gain long term)}
\item{\code{"DIVCGS"}}{ Dividend (capital gain short term)}
\item{\code{"DIVFEE"}}{ Dividend fee}
\item{\code{"DIVFT"}}{ Dividend adjusted (Foreign Tax Withheld)}
\item{\code{"DIVNRA"}}{ Dividend adjusted (NRA Withheld)}
\item{\code{"DIVROC"}}{ Dividend return of capital}
\item{\code{"DIVTW"}}{ Dividend adjusted (Tefra Withheld)}
\item{\code{"DIVTXEX"}}{ Dividend (tax exempt)}
\item{\code{"INT"}}{ Interest (credit/margin)}
\item{\code{"INTNRA"}}{ Interest adjusted (NRA Withheld)}
\item{\code{"INTTW"}}{ Interest adjusted (Tefra Withheld)}
\item{\code{"JNL"}}{ Journal entry}
\item{\code{"JNLC"}}{ Journal entry (cash)}
\item{\code{"JNLS"}}{ Journal entry (stock)}
\item{\code{"MA"}}{ Merger/Acquisition}
\item{\code{"NC"}}{ Name change}
\item{\code{"OPASN"}}{ Option assignment}
\item{\code{"OPEXP"}}{ Option expiration}
\item{\code{"OPXRC"}}{ Option exercise}
\item{\code{"PTC"}}{ Pass Thru Charge}
\item{\code{"PTR"}}{ Pass Thru Rebate}
\item{\code{"REORG"}}{ Reorg CA}
\item{\code{"SC"}}{ Symbol change}
\item{\code{"SSO"}}{ Stock spinoff}
\item{\code{"SSP"}}{ Stock split}
}
}
\examples{
account_activities(activity_type = "FILL")
}
\seealso{
Other Account: 
\code{\link{account_config}()},
\code{\link{account_portfolio}()},
\code{\link{account}()}
}
\concept{Account}
