% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{account_config}
\alias{account_config}
\title{Get Account Configurations}
\usage{
account_config(
  dtbp_check,
  fractional_trading,
  max_margin_multiplier,
  no_shorting,
  pdt_check,
  suspend_trade,
  trade_confirm_email,
  live = get_live()
)
}
\arguments{
\item{dtbp_check}{\code{(character)} both, entry, or exit. Controls Day Trading Margin Call (DTMC) checks. Default NULL (no change). Set to "default" to set all configuration options back to defaults. See Value section for defaults.}

\item{fractional_trading}{\code{(logical)}}

\item{max_margin_multiplier}{\code{(numeric)}}

\item{no_shorting}{\code{(logical)} If true, account becomes long-only mode. Default NULL (no change).}

\item{pdt_check}{\code{(character)} Not documented.}

\item{suspend_trade}{\code{(logical)} If true, new orders are blocked. Default NULL (no change).}

\item{trade_confirm_email}{\code{(character)} all or none. If none, emails for order fills are not sent. Default NULL (no change).}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
AccountConfigurations \code{(list)} \href{https://alpaca.markets/docs/api-documentation/api-v2/account-configuration/#accountconfigurations-entity}{AccountConfigurations}  object with the following parameters (defaults shown):
\itemize{
\item{\code{dtbp_check = "entry"}}
\item{\code{fractional_trading = TRUE}}
\item{\code{max_margin_multiplier = 4}}
\item{\code{no_shorting = FALSE}}
\item{\code{pdt_check = "entry"}}
\item{\code{suspend_trade = FALSE}}
\item{\code{trade_confirm_email = "all"}}
}
}
\description{
The account configuration API serves an important role in setting the way you want. See \href{https://alpaca.markets/docs/api-documentation/api-v2/account-configuration/}{Account Configurations} for details.
}
\examples{
\dontrun{
account_config(live = FALSE)
# Which is similar to:
account_config()
# For access to live accounts, you must submit as live = TRUE
account_config(live = TRUE)
account_config(dtbp_check = "both", no_shorting = TRUE)
account_config("default")
}
}
\seealso{
Other Account: 
\code{\link{account_activities}()},
\code{\link{account_portfolio}()},
\code{\link{account}()}
}
\concept{Account}
