% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{account_portfolio}
\alias{account_portfolio}
\title{Account Portfolio History}
\usage{
account_portfolio(
  period = NULL,
  timeframe = NULL,
  date_end = NULL,
  extended_hours = FALSE,
  live = get_live()
)
}
\arguments{
\item{period}{\code{(character/Duration/Period)} \emph{Optional} The period of time previous starting at \code{date_end} in \code{number + unit} format over which data will be provided. Such as one day \code{"1D"}.
\itemize{
\item{\code{`"D"`}}{ day }
\item{\code{`"W"`}}{ week }
\item{\code{`"M"`}}{ month }
\item{\code{`"A/Y"`}}{ year }
}
Defaults to \code{"1M"}. Accepts any number for the multiplier. Non case-sensitive. Also accepts \code{\link[lubridate]{period} & \link[lubridate]{Duration}} objects.}

\item{timeframe}{\code{(character)} \emph{Optional} The timeframe of the returned data in \code{"MT"} format where \code{M} is a multiplier and \code{T} is the timeframe from one of the following (non case-sensitive):
#' \itemize{
\item{\code{'m'}/\code{'min'}/\code{'minute'}}{ (\code{multiplier} can be \code{1}/\code{5}/\code{15})}
\item{\code{'h'}/\code{'hour'}}{ \code{multiplier} will be \code{1}.}
\item{\code{'d'}/\code{'day'}}{ (\code{multiplier} will be \code{1})}
}
If omitted, "\verb{1Min}" for less than 7 days period, "\verb{15Min}" for less than 30 days, or otherwise \code{"1D"}.}

\item{date_end}{\code{(Date/POSIXlt/Datetime(POSIXct)/character)} \emph{Optional} See Details for formatting guidelines. The end date of the returned data. Defaults to the current market date (rolls over at the market open if \code{extended_hours = FALSE}, otherwise at 7am ET). \emph{Note} that when \verb{period = 1D} and \code{date_end} is supplied, the API returns no data.}

\item{extended_hours}{\code{(logical)} \emph{Optional} If true, include extended hours in the result. This is effective only for timeframes less than 1D. Default \code{FALSE}}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
A \href{https://alpaca.markets/docs/api-documentation/api-v2/portfolio-history/#portfoliohistory-entity}{PortfolioHistory} \code{(tibble)} Object with R amenable data types. \emph{Note} that base value & timeframe can be accessed by calling \code{attr(out, "info")} on the returned tibble.
\itemize{
\item{\code{timestamp}}{ \code{(Date/POSIXlt/Datetime(POSIXct)/character)} time of each data element, with the first element being the beginning of the time window.}
\item{\code{equity}}{ \code{(numeric)} equity value of the account in dollar amount as of the end of each time window}
\item{\code{profit_loss}}{ \code{(numeric)} profit/loss in dollar from the base value}
\item{\code{profit_loss_pct}}{ \code{(numeric)} profit/loss in percentage from the base value}
\item{\code{base_value}}{ \code{(numeric)} basis in dollar of the profit loss calculation}
\item{\code{timeframe}}{ \code{(character)} time window size of each data element}
}
}
\description{
The \href{https://alpaca.markets/docs/api-documentation/api-v2/portfolio-history/}{portfolio history endpoint} returns a timeseries of equity and profit/loss summaries for the account over a \code{period} previous aggregated by a given \code{timeframe}.
}
\details{
All \code{(Date/POSIXlt)} will parse correctly if in \code{YYYY-MM-DD} \href{https://www.iso.org/iso-8601-date-and-time-format.html}{RFC 3339} format or \code{(Datetime/POSIXct)}, \verb{YYYY-MM-DD HH:MM} \href{https://www.iso.org/iso-8601-date-and-time-format.html}{ISO8601} format. Other formats will often work, but are not guaranteed to parse correctly.
}
\examples{
# Get the previous year's returns for the paper account
account_portfolio("1y")
# Get portfolio history when the COVID-19 pandemic overtook the US
account_portfolio("3m", date_end = "2020-05-20")
}
\seealso{
Other Account: 
\code{\link{account_activities}()},
\code{\link{account_config}()},
\code{\link{account}()}
}
\concept{Account}
