% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{add_row.tsymble_ts}
\alias{add_row.tsymble_ts}
\title{Add a row to a tsybble}
\usage{
add_row.tsymble_ts(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{Data frame to append to.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>
Name-value pairs, passed on to \code{\link[tibble:tibble]{tibble()}}. Values can be defined
only for columns that already exist in \code{.data} and unset columns will get an
\code{NA} value.}

\item{.before}{One-based row index where to add the new rows,
default: after last row.}

\item{.after}{One-based row index where to add the new rows,
default: after last row.}
}
\description{
\lifecycle{experimental}
Adds a row to a tsybble
}
\examples{
# add_row ---------------------------------
df <- tibble(x = 1:3, y = 3:1)

df \%>\% add_row(x = 4, y = 0)

# You can specify where to add the new rows
df \%>\% add_row(x = 4, y = 0, .before = 2)

# You can supply vectors, to add multiple rows (this isn't
# recommended because it's a bit hard to read)
df \%>\% add_row(x = 4:5, y = 0:-1)

# Use tibble_row() to add one row only
df \%>\% add_row(tibble_row(x = 4, y = 0))
try(df \%>\% add_row(tibble_row(x = 4:5, y = 0:-1)))

# Absent variables get missing values
df \%>\% add_row(x = 4)

# You can't create new variables
try(df \%>\% add_row(z = 10))
}
\seealso{
Other addition: 
\code{\link[tibble]{add_column}()}
}
\keyword{internal}
