% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlpacaforR-package.R
\docType{package}
\name{AlpacaforR-package}
\alias{AlpacaforR}
\alias{AlpacaforR-package}
\title{AlpacaforR: Trade with Alpaca using R}
\description{
Interact with the Alpaca API and the subset of endpoints on the Polygon API` available to Alpaca members. All Alpaca endpoints are accessible with AlpacaforR: Account, Assets, Calendar, Clock, Market-Data, Orders, Polygon, Positions, Watchlist and Streaming. Get asset, account and order info, market calendar info and historical market data. Submit, cancel and replace simple and complex orders, see market open / closing information, and much more. Build and test strategies with a paper account, then implement them live with an Alpaca Brokerage account. If you'd like to provide feedback, submit an issue, express thanks, or donate to support the project, please do so on Github: github.com/yogat3ch/AlpacaforR
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yogat3ch.github.io/AlpacaforR}
  \item Report bugs at \url{https://github.com/yogat3ch/AlpacaforR/issues}
}

}
\author{
\strong{Maintainer}: Holsenbeck Stephen \email{sholsen@alumni.emory.edu} (Primary Maintainer) [copyright holder]

Authors:
\itemize{
  \item Jagger Villalobos \email{jagger.vill@gmail.com} (Original Creator) [copyright holder]
}

}
\keyword{internal}
