% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assets.R
\name{asset_transform}
\alias{asset_transform}
\title{asset_transform}
\usage{
asset_transform(asts)
}
\arguments{
\item{asts}{\code{(response)}}
}
\value{
\code{(tibble)}
}
\description{
Transform API response to asset tibble
}
\keyword{Internal}
