% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assets.R
\name{assets}
\alias{assets}
\alias{get_assets}
\title{Get Assets}
\usage{
assets(ticker_id = NULL, v = 2)

get_assets(ticker_id = NULL, v = 2)
}
\arguments{
\item{ticker_id}{\code{(character)} Find asset details by symbol or ID. If NULL (the default), a list of assets is returned}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
Asset \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/assets/#asset-entity}{Asset} Object or array of Asset objects with the following parameters:
\itemize{
\item{\code{id}}{ \code{(character)} Asset ID as a string.}
\item{\code{class}}{ \code{(character)} us_equity as a string.}
\item{\code{exchange}}{ \code{(character)} AMEX, ARCA, BATS, NYSE, NASDAQ or NYSEARCA as a string.}
\item{\code{symbol}}{ \code{(character)} Stock symbol as a string.}
\item{\code{name}}{ \code{(character)} Full company name}
\item{\code{status}}{ \code{(character)} active or inactive as a string.}
\item{\code{tradable}}{ \code{(logical)} Asset is tradable on Alpaca or not.}
\item{\code{marginable}}{ \code{(logical)} Asset is marginable on Alpaca or not.}
\item{\code{shortable}}{ \code{(logical)} Asset is shortable on Alpaca or not.}
\item{\code{easy_to_borrow}}{ \code{(logical)} Asset is easy-to-borrow or not (filtering for \code{easy_to_borrow = TRUE} is the best way to check whether the name is currently available to short at Alpaca).}
}
}
\description{
The assets API serves as the master list of assets available for trade and data consumption from Alpaca. Assets are sorted by asset class, exchange and symbol. Some assets are only available for data consumption via Polygon, and are not tradable with Alpaca. These assets will be marked with the flag \code{tradable = FALSE}. See \href{https://alpaca.markets/docs/api-documentation/api-v2/assets#get-an-asset}{Assets: GET} for details.

\code{get_assets} is deprecated. Use \code{\link[AlpacaforR]{assets}} instead.
}
\examples{
# Get a tibble of all active assets: 
assets()
# Get a specific asset by symbol:
(AAPL <- assets("AAPL"))
# or by id:
(AAPL <- assets(AAPL$id))
get_assets()
}
\concept{Assets}
