% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{auth}
\alias{auth}
\title{Authorize an *Socket connection}
\usage{
auth(self, live)
}
\arguments{
\item{self}{\code{(websocket)} The websocket to authorize}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\description{
Authorizes the websocket connection passed as the first argument
}
