% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_bounds}
\alias{bars_bounds}
\title{Create API amenable boundaries based on user input for market_data}
\usage{
bars_bounds(...)
}
\arguments{
\item{from}{See \code{\link[AlpacaforR]{market_data}}}

\item{to}{See \code{\link[AlpacaforR]{market_data}}}

\item{after}{See \code{\link[AlpacaforR]{market_data}}}

\item{until}{See \code{\link[AlpacaforR]{market_data}}}

\item{v}{See \code{\link[AlpacaforR]{market_data}}}
}
\value{
\code{(list)} A list with two Date/Datetime items that are boundaries for the date range necessary to call the API with to retrieve the appropriate data from the specified API. Items will have names corresponding to the non-NULL \code{from/to/after/until} arguments in the calling environment or supplied to the function.
}
\description{
Internal function for fixing input date boundaries. The function is intended for use in an environment where its parameters already exist as it will use the variables from it's parent environment. See notes on specifying arguments in the \emph{Details} section for \code{\link[AlpacaforR]{bars_complete}}.
}
\seealso{
Other bars: 
\code{\link{bars_complete}()},
\code{\link{bars_expected}()},
\code{\link{bars_get}()},
\code{\link{bars_missing}()},
\code{\link{bars_transform}()},
\code{\link{bars_url}()}
}
\concept{bars}
\keyword{internal}
