% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_expected}
\alias{bars_expected}
\title{Return a Date/Datetime vector that the user expects}
\usage{
bars_expected(bars, ...)
}
\arguments{
\item{bars}{\code{(list)} (Required) Output from \code{\link[AlpacaforR]{bars_get}}}

\item{v}{See \code{\link[AlpacaforR]{market_data}}}

\item{.bounds}{See \code{\link[AlpacaforR]{bars_bounds}}}

\item{multiplier}{See \code{\link[AlpacaforR]{market_data}}}

\item{timeframe}{See \code{\link[AlpacaforR]{market_data}}}

\item{.tf_num}{Numeric rank of timeframe, calculated automatically if \code{timeframe} is passed or available in the parent env.}
}
\value{
\code{(Date/Datetime)} A vector of expected date/datetimes based on the \code{.bounds}, \code{timeframe} and \code{multiplier}
}
\description{
An internal function for use when param \code{full = T} in \code{\link[AlpacaforR]{market_data}} that creates a vector of expected Date/Datetimes that are expected from a given call to either v1 or v2 API. The function is intended for use in an environment where its parameters already exist. See notes on specifying arguments in the \emph{Details} section for \code{\link[AlpacaforR]{bars_complete}}.
}
\keyword{internal}
