% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Market_Data.R
\name{bars_span}
\alias{bars_span}
\title{memorize the span of data that the API returns each time}
\usage{
bars_span(x, ext, timeframe)
}
\arguments{
\item{x}{\code{(tsymble)} returned from API call}

\item{ext}{\code{(environment)} with cyclically modified variables in \code{bars_complete}}
}
\value{
\code{(\link[lubridate]{Interval})} of the time span returned by the API. Saved to the \code{ext} environment.
}
\description{
memorize the span of data that the API returns each time
}
\keyword{Internal}
