% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_url}
\alias{bars_url}
\title{Create URLs for market_data endpoints}
\usage{
bars_url(..., limit = NULL)
}
\arguments{
\item{limit}{\emph{v1 only} `See \code{\link[AlpacaforR]{market_data}}, NULL unless explicitly specified.}

\item{ticker}{See \code{\link[AlpacaforR]{market_data}}}

\item{.bounds}{Output from [AlpacaforR]{bars_bounds}. If the output is assigned  to \code{.bounds} in the calling environment this object will be used.}

\item{multiplier}{See \code{\link[AlpacaforR]{market_data}}}

\item{timeframe}{See \code{\link[AlpacaforR]{market_data}}}

\item{v}{See \code{\link[AlpacaforR]{market_data}}}
}
\value{
\code{(character)} URL to be called with \code{\link[AlpacaforR]{bars_get}}
}
\description{
Internal function for generating query urls. The function is intended for use in an environment where its parameters already exist. See notes on specifying arguments in the \emph{Details} section for \code{\link[AlpacaforR]{bars_complete}}.
}
\keyword{internal}
