% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{bind_rows}
\alias{bind_rows}
\title{bind_rows}
\usage{
bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dplyr:bind_rows]{dplyr::bind_rows}}
  \describe{
    \item{\code{}}{}
  }}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use
names if available, otherwise it will use positions.}
}
\value{
A data frame the same type as the first element of \code{...}.
}
\description{
Replace \code{dplyr::bind_rows} for use with \code{tsymble_ts} objects.
}
