% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calendar.R
\name{calendar}
\alias{calendar}
\alias{get_calendar}
\title{Get Market Calendar Times}
\usage{
calendar(from = NULL, to = NULL, v = 2)

get_calendar(from = NULL, to = NULL, version = "v2")
}
\arguments{
\item{from}{\code{(Date/Datetime/POSIXct/character)} in YYYY-MM-DD format. Starting date of request. Defaults to today. Calendar data goes back to 1970.}

\item{to}{\code{(Date/Datetime/POSIXct/character)} in YYYY-MM-DD format. Ending date of request. Default to today. Calendar data goes to 2029.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
Calendar \code{tibble} A \href{https://alpaca.markets/docs/api-documentation/api-v2/calendar/#calendar-entity}{Calendar} Object consisting of the following:
\itemize{
\item{\code{date}}{ \code{(character)} vector of trading days in the time range}
\item{\code{open}}{ \code{(character)} The time the market opens at on this date in hour:min format as a string.}
\item{\code{close}}{ \code{(character)} The time the market closes at on this date in hour:min format as a string.}
\item{\code{session_open}}{ \code{(character)} The start time of the session in hh:mm format}
\item{\code{session_close}}{ \code{(character)} The closing time of the session in hh:mm format.}
\item{\code{day}}{ \code{(Interval)} An interval of the market day \code{open \%--\% close}.}
\item{\code{session}}{ \code{(Interval)} An interval of the market session \code{session_open \%--\% session_close}.}
\item{\code{dow}}{ \code{(character)} Three letter abbreviation of the day of the week as an ordered factor with levels Sunday - Saturday. }
}
}
\description{
The calendar API serves the full list of market days from 1970 to 2029. It can also be queried by specifying a start and/or end time to narrow down the results. In addition to the dates, the response also contains the specific open and close times for the market days, taking into account early closures. See \href{https://alpaca.markets/docs/api-documentation/api-v2/calendar/}{Calendar} Endpoint for details.

\code{get_calendar} is Deprecated. Use \code{\link[AlpacaforR]{calendar}} instead.
}
\examples{
# Get all dates:
calendar()
# Get specific date range:
calendar(from = "2019-01-01", to = "2019-04-01")
get_calendar()
}
\seealso{
\link[lubridate]{Interval-class}
}
\concept{Calendar}
