% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Order function}
\usage{
cancel_order(ticker_id, live = FALSE)
}
\arguments{
\item{ticker_id}{The ticker symbol or the order id.}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\description{
Cancels any open order by either ticker or order id. If multiple open orders exist for one ticker, then the default is to cancel the most recent order.
}
\examples{
cancel_order(ticker_id = "AAPL")
cancel_order(ticker_id = "aapl")
Or you can instead cancel by the order_id:
orders <- get_orders(status="open", silent = TRUE)
cancel_order(ticker_id = orders$id[1])
}
