% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{channel_bars}
\alias{channel_bars}
\title{Determine the socket from the channel}
\usage{
channel_bars(channel)
}
\arguments{
\item{channel}{\code{(character)} Name of the channel}
}
\description{
Determine the socket from the channel
}
