% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{class_prefix}
\alias{class_prefix}
\title{Extract the socket name from class of object}
\usage{
class_prefix(self)
}
\arguments{
\item{self}{\code{(*Socket)} A socket object}
}
\value{
\code{(character)} the name of the socket (\code{"Alpaca"/"Polygon"})
}
\description{
\lifecycle{stable} Determines the socket from the class of the object
}
\keyword{Internal}
