% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positions.R
\name{close_position}
\alias{close_position}
\title{Close Positions by ticker, or close all (Deprecated)}
\usage{
close_position(ticker = NULL, live = FALSE, v = 2)
}
\arguments{
\item{ticker}{\code{(character)} \emph{required} the symbol of the ticker for which positions will be closed. Use "close_all" to close all positions.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
Position \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/#position-entity}{Position object} or array of Position objects of length 0 if no positions, otherwise with length of the number of positions, each with the following attributes:
\itemize{
\item{\code{asset_id}}{\code{(character)} Asset ID.}
\item{\code{symbol}}{\code{(character)} Symbol of the asset.}
\item{\code{exchange}}{\code{(character)} Exchange name of the asset.}
\item{\code{asset_class}}{\code{(character)} Asset class name.}
\item{\code{qty}}{\code{(integer)} The number of shares.}
\item{\code{avg_entry_price}}{\code{(numeric)} Average entry price of the position.}
\item{\code{side}}{\code{(character)} long/short exposure.}
\item{\code{market_value}}{\code{(numeric)} Total dollar amount of the position.}
\item{\code{cost_basis}}{\code{(numeric)} Total cost basis in dollar.}
\item{\code{unrealized_pl}}{\code{(numeric)} Unrealized profit/loss in dollar.}
\item{\code{unrealized_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{unrealized_intraday_pl}}{\code{(numeric)} Unrealized profit/loss in dollar for the day.}
\item{\code{unrealized_intraday_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{current_price}}{\code{(numeric)} Current asset price per share.}
\item{\code{lastday_price}}{\code{(numeric)} Last day's asset price per share.}
\item{\code{change_today}}{\code{(numeric)} Percent change from last day price (by a factor of 1).}
}
}
\description{
\code{close_position} is deprecated. Use \code{\link[AlpacaforR]{positions}} with \code{action = "close"} instead. The close_positions function liquidates all positions or those specified by \code{ticker}. Character values are returned as a string while integer values are returned as numeric. See \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/}{Positions} for details.
}
\examples{
close_position(ticker = "AAPL", live = FALSE)
close_position(ticker = "AAPL")
close_position(ticker = "AAPL", live = TRUE)
}
\seealso{
Other Positions: 
\code{\link{positions}()}
}
\concept{Positions}
