% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{console_msg}
\alias{console_msg}
\title{Send a \code{msg} to the console}
\usage{
console_msg(msg, opts)
}
\arguments{
\item{msg}{\code{(msg/tbl)} The \code{msg} tibble object returned from \href{../../AlpacaforR/html/msg.html}{\code{msg()}} which consists of the following:
\itemize{
\item{\code{ts}}{ \code{(POSIXct)} timestamp}
\item{\code{socket}}{ \code{(character)} the socket over which the message arrived}
\item{\code{channel}}{ \code{(character)} the channel over which the message arrived}
\item{\code{...}}{ \code{(varies)} all fields encoded in the event data JSON}
}}

\item{opts}{\code{(list)} The list of options from the *Socket object. Used to determine whether the object should be sent to console.}
}
\description{
Send a \code{msg} to the console
}
\keyword{Internal}
