% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{console_prefix}
\alias{console_prefix}
\title{The socket name formatted for display on the console}
\usage{
console_prefix(msg)
}
\arguments{
\item{msg}{\code{(msg/tbl)} The \code{msg} tibble object returned from \href{../../AlpacaforR/html/msg.html}{\code{msg()}} which consists of the following:
\itemize{
\item{\code{ts}}{ \code{(POSIXct)} timestamp}
\item{\code{socket}}{ \code{(character)} the socket over which the message arrived}
\item{\code{channel}}{ \code{(character)} the channel over which the message arrived}
\item{\code{...}}{ \code{(varies)} all fields encoded in the event data JSON}
}}
}
\value{
\code{(character)} prefix for console display
}
\description{
Use colors specific to each websocket for easy visual discernment of which socket is reporting to the console
}
\keyword{Internal}
