% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{create_watchlist}
\alias{create_watchlist}
\title{Create Watchlist function}
\usage{
create_watchlist(name = NULL, tickers = c(NULL), live = FALSE)
}
\arguments{
\item{name}{arbitrary name string, up to 64 characters. <string REQUIRED>}

\item{tickers}{Set of symbols. <vector of string>}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
An array of a watchlist object.
}
\description{
Creates a new watchlist with initial set of assets.
}
\examples{
create_watchlist(name = "test", tickers = c("AAPL","WMT"), live = FALSE)
}
