% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R, R/alpaca4R.R
\name{delete_from_watchlist}
\alias{delete_from_watchlist}
\title{Delete from Watchlist (Deprecated)}
\usage{
delete_from_watchlist(watchlist_id = NULL, ticker = NULL, live = FALSE)

delete_from_watchlist(watchlist_id = NULL, ticker = NULL, live = FALSE)
}
\arguments{
\item{watchlist_id}{Watchlist ID <string>}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{tickers}{Symbol name to remove from the watchlist content. <string>}
}
\description{
Deprecated, use \code{\link[AlpacaforR]{watchlist}} with \code{action = "delete"} instead. Delete one entry for an asset by symbol name.

Delete one entry for an asset by symbol name.
}
\examples{
\dontrun{
delete_from_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", ticker = "AAPL", live = FALSE)
}
delete_from_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", ticker = "AAPL", live = FALSE)
}
\seealso{
watchlist_update

Other Watchlist: 
\code{\link{add_to_watchlist}()},
\code{\link{delete_watchlist}()},
\code{\link{watchlist_update}()},
\code{\link{watchlist}()}
}
\concept{Watchlist}
