% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{dplyr_col_modify.tsymble_ts}
\alias{dplyr_col_modify.tsymble_ts}
\title{dplyr_col_modify.tsymble_ts}
\usage{
\method{dplyr_col_modify}{tsymble_ts}(data, cols)
}
\arguments{
\item{data}{A tibble. We use tibbles because they avoid some inconsistent
subset-assignment use cases}

\item{cols}{A named list used modify columns. A \code{NULL} value should remove
an existing column.}
}
\description{
An S3 method for modifying a \code{tsymble_ts} object
}
