% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{dplyr_reconstruct.tsymble_ts}
\alias{dplyr_reconstruct.tsymble_ts}
\title{dplyr_reconstruct.tsymble_ts}
\usage{
\method{dplyr_reconstruct}{tsymble_ts}(data, template)
}
\arguments{
\item{data}{A tibble. We use tibbles because they avoid some inconsistent
subset-assignment use cases}

\item{template}{Template to use for restoring attributes}
}
\description{
An S3 method for reconstructing a \code{tsymble_ts} object
}
