% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{dplyr_row_slice.grouped_tsymble}
\alias{dplyr_row_slice.grouped_tsymble}
\title{dplyr_row_slice.grouped_tsymble}
\usage{
\method{dplyr_row_slice}{grouped_tsymble}(data, i, ...)
}
\arguments{
\item{data}{A tibble. We use tibbles because they avoid some inconsistent
subset-assignment use cases}

\item{i}{A numeric or logical vector that indexes the rows of \code{.data}.}

\item{...}{
  Arguments passed on to \code{\link[dplyr:dplyr_extending]{dplyr::dplyr_row_slice}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
An S3 method for slicing a \code{tsymble_ts} object
}
