% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{dplyr_row_slice.tsymble_ts}
\alias{dplyr_row_slice.tsymble_ts}
\title{distinct.tsymble_ts <- function(.data, ...) {
dplyr::distinct(dplyr::as_tibble(.data), ...)
#TODO Didn't we already redefine this?
}}
\usage{
dplyr_row_slice.tsymble_ts(data, i, ..., preserve = FALSE)
}
\description{
#' @export
dplyr_row_slice.tbl_ts <- function(data, i, ..., preserve = FALSE) {
tbl <- dplyr::as_tibble(data)
loc_df <- dplyr::summarise(tbl, !!".loc" := rlang::list2(i))
ascending <- all(purrr::map_lgl(loc_df[\link{".loc"}], validate_order))
res <- dplyr::dplyr_row_slice(tbl, i, ..., preserve = preserve)
if (preserve) {
update_meta2(res, data, ordered = ascending, interval = tsibble::interval(data))
} else {
update_meta(res, data, ordered = ascending, interval = tsibble::interval(data))
}
}
}
\section{either key or index is present in \code{cols}}{

}

\section{suggests that the operations are done on these variables}{

}

\section{validate = TRUE to check if tsibble still holds}{

}

