% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{event_data}
\alias{event_data}
\title{Extract event data}
\usage{
event_data(., msg)
}
\arguments{
\item{.}{\code{(event$data)} object returned by the websocket}

\item{msg}{\code{(msg/tbl)} The \code{msg} tibble object returned from \href{../../AlpacaforR/html/msg.html}{\code{msg()}} which consists of the following:
\itemize{
\item{\code{ts}}{ \code{(POSIXct)} timestamp}
\item{\code{socket}}{ \code{(character)} the socket over which the message arrived}
\item{\code{channel}}{ \code{(character)} the channel over which the message arrived}
\item{\code{...}}{ \code{(varies)} all fields encoded in the event data JSON}
}}
}
\value{
\code{(event data/data.frame)} Returns event data that is typically a \code{tbl/data.frame}
}
\description{
Extract message data from \code{event} list object
}
\keyword{Internal}
