% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fetch_vars}
\alias{fetch_vars}
\title{fetch variables}
\usage{
fetch_vars(
  .vn,
  e = NULL,
  cenv = rlang::caller_env(),
  penv = parent.env(rlang::caller_env()),
  sf = rev(sys.frames())
)
}
\arguments{
\item{.vn}{named vector of variable names ie for a variable \code{d} \code{.vn = c(d = "d")}}

\item{e}{The list made from ellipsis if parameters are input as ellipsis argument}

\item{cenv}{The caller environment, stored automatically}

\item{penv}{The parent of the caller environment, also stored automatically}

\item{sf}{The system frames for searching if not found in previous two environments.}
}
\description{
Fetches variables from environment of containing function for use in internal function
}
\keyword{internal}
