% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Market_Data.R
\name{gap_check}
\alias{gap_check}
\title{ensure the gap returned from \code{gap_identify} is valid}
\usage{
gap_check(x, ext)
}
\arguments{
\item{x}{\code{(tibble)} gap data from \code{.from} and \code{.to} columns}

\item{ext}{\code{(environment)} with cyclically modified variables in \code{bars_complete}}
}
\value{
\code{(logical)} indicating whether \code{gap_identify} should progress to the next form of gap identification because the gap type provided is invalid. \code{TRUE} means all gaps of type have been identified, indicated by invalid dates in \code{.from} & \code{to} or the gap object being identical to the previous round. \code{FALSE} means gap exists, and proceed filling gaps.
}
\description{
ensure the gap returned from \code{gap_identify} is valid
}
\keyword{Internal}
