% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_account_activities}
\alias{get_account_activities}
\title{Get Account function}
\usage{
get_account_activities(
  activity_type = c(NULL),
  date = NULL,
  until = NULL,
  after = NULL,
  direction = "desc",
  page_size = 50,
  page_token = NULL,
  live = FALSE
)
}
\arguments{
\item{activity_type}{The activity type you want to view entries for. A list of valid activity types can be found in the examples below. <string>}

\item{date}{The date for which you want to see activities. <string timestamp>}

\item{until}{The response will contain only activities submitted before this date. (Cannot be used with date.) <string timestamp>}

\item{after}{The response will contain only activities submitted after this date. (Cannot be used with date.) <string timestamp>}

\item{direction}{asc or desc, default is desc. <string>}

\item{page_size}{The maximum number of entries to return in the response. <int>}

\item{page_token}{The ID of the end of your current page of results. <string>}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
"id" Account ID as a string.

"status" Account Status as a string.

"currency" USD as a string.

"buying_power" Tradable buying power as a string.

"cash" Cash balance as a string.

"cash_withdrawable"  Withdrawable cash amount as a string.

"portfolio_value"  Total value of cash + holding positions as a string.

"trading_blocked"  If true, the account is not allowed to place orders as a boolean.

"transfers_blocked"  If true, the account is not allowed to request money transfers as a boolean.

"account_blocked"  If true, the account activity by user is prohibited as a boolean.

"created_at"  Timestamap this account was created at as a string.
}
\description{
The accounts API serves important information related to an account, including account status, funds available for trade, funds available for withdrawl, and various flags relevant to an account’s ability to trade.
}
\examples{
get_account_activities(activity_type = "FILL")
Activity Types:
FILL: Order fills (both partial and full fills)
TRANS: Cash transactions (both CSD and CSR)
MISC: Miscellaneous or rarely used activity types (All types except those in TRANS, DIV, or FILL)
ACATC: ACATS IN/OUT (Cash)
ACATS: ACATS IN/OUT (Securities)
CSD: Cash disbursement(+)
CSR: Cash receipt(-)
DIV: Dividends
DIVCGL: Dividend (capital gain long term)
DIVCGS: Dividend (capital gain short term)
DIVFEE: Dividend fee
DIVFT: Dividend adjusted (Foreign Tax Withheld)
DIVNRA: Dividend adjusted (NRA Withheld)
DIVROC: Dividend return of capital
DIVTW: Dividend adjusted (Tefra Withheld)
DIVTXEX: Dividend (tax exempt)
INT: Interest (credit/margin)
INTNRA Interest adjusted (NRA Withheld)
INTTW: Interest adjusted (Tefra Withheld)
JNL: Journal entry
JNLC: Journal entry (cash)
JNLS: Journal entry (stock)
MA: Merger/Acquisition
NC: Name change
OPASN: Option assignment
OPEXP: Option expiration
OPXRC: Option exercise
PTC: Pass Thru Charge
PTR: Pass Thru Rebate
REORG: Reorg CA
SC: Symbol change
SSO: Stock spinoff
SSP: Stock split
}
