% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_bars}
\alias{get_bars}
\title{Get Bars function}
\usage{
get_bars(url_data, headers, ticker, from)
}
\arguments{
\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}

\item{ticker}{The stock or stocks (in vector format) that you want.}

\item{from}{The starting date for the pricing data}

\item{url}{URL to your user live or paper API.}
}
\value{
"t" the beginning time of this bar as a Unix epoch in seconds as a integer.

"o" open price as a numberic object.

"h" high price as a numberic object.

"l" low price as a numberic object.

"c" close price as a numberic object.

"v" volume as a numberic object.
}
\description{
The bars API provides time-aggregated price and volume data for a single stock or multiple.
}
\examples{
get_bars(url_data = url_data, headers = headers, ticker = "INTC", from = "2019-03-20")
}
