% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_calendar}
\alias{get_calendar}
\title{Get Calendar function}
\usage{
get_calendar(url, headers)
}
\arguments{
\item{url}{URL to your user live or paper API.}

\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}
}
\value{
"date" Date string. in date format as a string.

"open" The time the market opens at on this date in hour:min format as a string.

"close" The time the market closes at on this date in hour:min format as a string.
}
\description{
The calendar API serves the full list of market days from 1970 to 2029. It can also be queried by specifying a start and/or end time to narrow down the results. In addition to the dates, the response also contains the specific open and close times for the market days, taking into account early closures.
}
\examples{
get_calendar(url = url, headers = headers)
}
