% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_clock}
\alias{get_clock}
\title{Get Clock function}
\usage{
get_clock()
}
\value{
"timestamp" Current timestamp as a string.

"is_open" Whether or not the market is open as a boolean.

"next_open" Next market open timestamp as a string.

"next_close" Next market close timestamp as a string.
}
\description{
The clock API serves the current market timestamp, whether or not the market is currently open, as well as the times of the next market open and close.
}
\examples{
get_clock()
}
