% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_config}
\alias{get_config}
\title{Get Account Configurations function}
\usage{
get_config(live = FALSE)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
"dtbp_check" both, entry, or exit. Controls Day Trading Margin Call (DTMC) checks.

"trade_confirm_email" all or none. If none, emails for order fills are not sent.

"suspend_trade" If true, new orders are blocked.

"no_shorting" If true, account becomes long-only mode.
}
\description{
The account configuration API serves an important role in setting the way you want.......
}
\examples{
get_config(live = FALSE)
Which is similar to:
get_config()
For access to live accounts, you must submit as live = TRUE
get_config(live = TRUE)
}
