% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_headers}
\alias{get_headers}
\title{return headers for API calls}
\usage{
get_headers(live = get_live())
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to NULL, so it will use the key variables set by the user for their respective paper account. Set live = TRUE to find your live key credentials.}
}
\value{
The correct headers that will be sent in the API request.
}
\description{
return headers for API calls
}
\keyword{internal}
