% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_key}
\alias{get_key}
\title{Get an API key from Environmental Variables}
\usage{
get_key(live = get_live(), api = "a")
}
\arguments{
\item{live}{\code{(logical)} Alpaca live setting}

\item{api}{\code{(character)} \code{"a"/"p"} for Alpaca/Polygon respectively}
}
\value{
\code{(character)} the requested key
}
\description{
\verb{APCA-[LIVE/PAPER]-KEY} or \code{POLYGON-KEY}s are character vectors set in the Renviron file, see \link[AlpacaforR]{firstrun} for details on setting up these credentials.
}
