% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_live}
\alias{get_live}
\title{Get the Alpaca API default live setting}
\usage{
get_live()
}
\value{
\code{(logical)} Alpaca API live setting
}
\description{
\code{live} is a logical argument to many package functions that indicates whether to use the live account (\code{TRUE}) or paper account (\code{FALSE}). The default is set by environmental variable, see \link[AlpacaforR]{firstrun} for details on setting up the default live setting.
}
