% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_meta}
\alias{get_meta}
\title{Get Polygon Meta Data}
\usage{
get_meta(ticker = NULL, endpoint = NULL, perpage = NULL, version = "v1")
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}

\item{endpoint}{Select either company for a company profile, analysts for all kinds of analyst estimates, dividends to view historic and upcoming dividends, earnings for historic and current earning details , or news for news updates from CNBC, Seeking Alpha, etc.}

\item{perpage}{This is only used if "news" was your selected endpoint. How many articles do you want to see perpage?}

\item{version}{The current version for API. Defaults to v1 if no v2 available.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides more color on your stock through its available meta data endpoints from Polygon. These endpoints are company, analysts, dividends, earnings, and news.
}
\examples{
Getting default meta for AMZN: 
get_meta(ticker = "AMZN")
Getting news information on AMZN: 
get_meta(ticker = "AMZN", endpoint = "news", perpage = 100)
}
