% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_orders}
\alias{get_orders}
\title{Get Orders function}
\usage{
get_orders(
  ticker = NULL,
  status = "open",
  limit = NULL,
  after = NULL,
  until = NULL,
  direction = "desc",
  silent = FALSE,
  live = FALSE,
  version = "v2"
)
}
\arguments{
\item{ticker}{\code{(character)} Specify which symbol you want to call by inserting ticker as a string.}

\item{status}{\code{(character)} Order status to be queried. \code{open}, \code{closed} or \code{all}. Defaults to open as a string.}

\item{limit}{\code{(integer)} The maximum number of orders in response. Defaults to 50 and max is 500.}

\item{after}{\code{(Date/character)} Date in YYYY-MM-DD \href{https://www.iso.org/iso-8601-date-and-time-format.html}{(ISO8601 Format)} The response will include only orders submitted \emph{after} this date exclusive as a timestamp object.}

\item{until}{\code{(Date/character)} Date in YYYY-MM-DD \href{https://www.iso.org/iso-8601-date-and-time-format.html}{(ISO8601 Format)} The response will include only orders submitted \emph{before} this date exclusive as a timestamp object.}

\item{direction}{\code{(character)} The chronological order of response based on the submission time. \code{asc} or \code{desc}. Defaults to \code{desc}.}

\item{silent}{\code{(logical)} TRUE / FALSE on if you want the "no orders to cancel" message to print to the console. Default to FALSE.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
"id" order id.

"client_order_id" client unique order id.

"created_at" nullable When the order was created.

"updated_at" nullable When the order was updated.

"submitted_at" nullable When the order was submitted.

"filled_at" nullable When the order was filled.

"expired_at" nullable When the order was expired.

"canceled_at" nullable When the order was canceled.

"asset_id" asset ID.

"symbol" Asset symbol.

"exchange" Asset exchange.

"asset_class" Asset class.

"qty" Ordered quantity.

"filled_qty" Filled quantity.

"type" Valid values: market, limit, stop, stop_limit.

"side" Valid values: buy, sell.

"time_in_force" time in force.

"limit_price" Limit price.

"stop_price" Stop price.

"status" Status of the order.
}
\description{
The orders API allows a user to monitor, place and cancel their orders with Alpaca. Times are returned as yyyy-mm-dd hh-mm-ss POSIXct, quantity and price as numeric, and all others as a string.
}
\examples{
get_orders(live = FALSE, version = "v2")
get_orders(status = "all", version = "v2")
# For a specific ticker:
get_orders(ticker = "AAPL", status = "all", version = "v2")
}
