% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R, R/alpaca4R.R
\name{get_poly_last_price}
\alias{get_poly_last_price}
\title{Get Polygon Last Price (Deprecated)}
\usage{
get_poly_last_price(ticker = NULL)

get_poly_last_price(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.

A list object containing all information the API responds with.
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides the last listed price from Polygon. A list is returned with values such as the last price, last size, last exchange, and last timestamp.

This function provides the last listed price from Polygon. A list is returned with values such as the last price, last size, last exchange, and last timestamp.
}
\examples{
# Getting the last listed price for AMZN: 
get_poly_last_price("AMZN")
# Getting the last listed price for AMZN: 
get_poly_last_price("AMZN")
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()},
\code{\link{polygon}()}
}
\concept{Polygon}
