% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{get_poly_last_trade}
\alias{get_poly_last_trade}
\title{Get Polygon Last Trade (Deprecated)}
\usage{
get_poly_last_trade(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides the last listed trade from Polygon for the ticker specified. A list is returned with values such as the last bid / ask price, bid / ask size, bid / ask exchange, and last trade timestamp.
}
\examples{
# Getting the last listed trade for AMZN: 
get_poly_last_trade("AMZN")
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()},
\code{\link{polygon}()}
}
\concept{Polygon}
