% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{get_poly_ohlc}
\alias{get_poly_ohlc}
\title{Get Polygon Daily OHLCV & After Hours (Deprecated)}
\usage{
get_poly_ohlc(ticker = NULL, date = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}

\item{date}{Specify the date for which you are requesting.}
}
\value{
A list object containing all information the API responds with.
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides the last OHLCV data, including after hours data from Polygon for the ticker specified.
}
\examples{
# Getting the last listed trade for AMZN: 
get_poly_ohlc("AMZN", date = "2019-03-20")
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()},
\code{\link{polygon}()}
}
\concept{Polygon}
