% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_prev_dayclose}
\alias{get_poly_prev_dayclose}
\title{Get Polygon Previous Day Close for Ticker}
\usage{
get_poly_prev_dayclose(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides the Previous Day close data from Polygon for the ticker specified.
}
\examples{
# Getting the last listed trade for AMZN: 
get_poly_prev_dayclose("AMZN")
}
