% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_pro}
\alias{get_pro}
\title{Is the package configured to use Alpaca Unlimited Subscription}
\usage{
get_pro()
}
\value{
\code{(logical)}
}
\description{
Returns \code{TRUE} if APCA-PRO environmental variable is set to \code{TRUE}. Set this variable to \code{TRUE} if subscribed to \href{https://alpaca.markets/docs/api-documentation/api-v2/market-data/alpaca-data-api-v2/#subscription-plans}{Alpaca's Unlimited Plan}. See \link{firstrun} for details on setting this variable.
}
