% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{get_query}
\alias{get_query}
\title{get_query}
\usage{
get_query(x)
}
\arguments{
\item{x}{\code{(tsybble)}}
}
\value{
\code{(character)}
}
\description{
\lifecycle{experimental}
Retrieve the query metadata from a \code{tsymble} returned from \link[AlpacaforR]{market_data}.
}
