% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_secret}
\alias{get_secret}
\title{Get an API secret from Environmental Variables}
\usage{
get_secret(live = get_live())
}
\arguments{
\item{live}{\code{(logical)} Alpaca live setting}
}
\value{
\code{(character)} the requested secret
}
\description{
\verb{APCA-[LIVE/PAPER]-SECRET} are character vectors set in the Renviron file, see \link[AlpacaforR]{firstrun} for details on setting up these credentials.
}
