% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{get_sym}
\alias{get_sym}
\title{Retrieve the ticker symbol}
\usage{
get_sym(x)
}
\arguments{
\item{x}{\code{(tsybble)}}
}
\value{
\code{(character)}
}
\description{
\lifecycle{experimental}
Retrieve the ticker symbol from a \code{tsymble} returned from market_data.
}
