% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Market_Data.R
\name{get_tibble}
\alias{get_tibble}
\title{Find object with \code{nm}}
\usage{
get_tibble(x, nm = "time", depth = FALSE)
}
\arguments{
\item{x}{\code{(list)}}

\item{nm}{\code{(character)} The name to recursively search for}

\item{depth}{\code{(logical)} Whether return the object (\code{FALSE}) \emph{default}, or an integer for the depth at which the object was found (\code{TRUE})}
}
\value{
\code{(numeric/tibble)}
}
\description{
Recurse into the top level list until an object that has a name matching \code{nm} is found
}
\keyword{Internal}
