% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_url_poly}
\alias{get_url_poly}
\title{Return the polygon URL}
\usage{
get_url_poly()
}
\value{
The correct URL for Polygon's API.
}
\description{
Get Polygon's URL for the Server Request that is sent to interact with the API.
}
\keyword{internal}
