% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_watchlist_ids}
\alias{get_watchlist_ids}
\title{Get Watchlist IDs function}
\usage{
get_watchlist_ids(live = FALSE)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
An array of watchlist objects.
}
\description{
Returns a list of watchlists and the IDs registered under the account.
}
\examples{
get_watchlist_ids(live = FALSE)
}
