% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{grapes-or-z-or-grapes}
\alias{grapes-or-z-or-grapes}
\alias{\%|z|\%}
\title{Default value for zero length variable}
\usage{
lhs \%|z|\% rhs
}
\arguments{
\item{x}{(lhs) If \code{x} is length zero,}

\item{y}{(rhs)  will return \code{y}; otherwise returns \code{x}}
}
\description{
This infix replaces zero length variables with a default value
}
\examples{
numeric(0) \%|z|\% NA_real_
}
