% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_id}
\alias{is_id}
\title{Check if value provided is an Alpaca ID}
\usage{
is_id(.)
}
\arguments{
\item{.}{\code{(character)}}
}
\value{
\code{logical} indicating whether the object is an id
}
\description{
for use in functions that accept \verb{*_id}
}
\keyword{internal}
