% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{is_legit}
\alias{is_legit}
\title{Does x contain information?}
\usage{
is_legit(x)
}
\arguments{
\item{x}{input}
}
\value{
\code{(logical)}
}
\description{
Returns TRUE if x is non-empty, non-NA, and non-null
}
