% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{is_ws_bars}
\alias{is_ws_bars}
\title{Is a message a bar object}
\usage{
is_ws_bars(msg)
}
\arguments{
\item{msg}{\code{(tibble)} the message}
}
\value{
\code{(logical)}
}
\description{
Websockets receive many types of messages, this function uses regex to determine if the message is a bars object
}
