% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{lgl_pcdnce}
\alias{lgl_pcdnce}
\title{Determine logical precedence}
\usage{
lgl_pcdnce(...)
}
\arguments{
\item{...}{logical values in order of precedence}
}
\value{
\code{(logical)}
}
\description{
return a single logical from a series of logical values in order of precedence (specific to default)
}
\keyword{Internal}
