% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{log_msg}
\alias{log_msg}
\title{message logging}
\usage{
log_msg(msg, opts, private)
}
\arguments{
\item{msg}{\code{(msg/tbl)} The \code{msg} tibble object returned from \href{../../AlpacaforR/html/msg.html}{\code{msg()}} which consists of the following:
\itemize{
\item{\code{ts}}{ \code{(POSIXct)} timestamp}
\item{\code{socket}}{ \code{(character)} the socket over which the message arrived}
\item{\code{channel}}{ \code{(character)} the channel over which the message arrived}
\item{\code{...}}{ \code{(varies)} all fields encoded in the event data JSON}
}}

\item{opts}{\code{(list)} The list of options from the *Socket object. Used to determine whether the object should be sent to console.}

\item{private}{\code{(environment)} The \verb{*Socket} private environment}
}
\description{
Logs message based on conditions supplied as options to *Socket object.
}
\keyword{Internal}
