% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{match_letters}
\alias{match_letters}
\title{Match the first \code{n} letters to supplied arguments}
\usage{
match_letters(
  x,
  ...,
  n = 1,
  multiple = FALSE,
  ignore.case = FALSE,
  capitalize = FALSE
)
}
\arguments{
\item{x}{\code{(character)} to match on}

\item{...}{\code{(character)} vectors to match against}

\item{n}{\code{(numeric)} how many characters of \code{x} to use in matching. Set to \code{NULL} to use all}

\item{multiple}{\code{(logical)} are multiple matches allowed? If \code{FALSE} (Default) only the first match is returned.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}

\item{capitalize}{\code{(logical)} whether to capitalize the result}
}
\value{
\code{(character)} vector of matches
}
\description{
Case insensitive matching of argument to possibilities provided in ellipsis.
}
