% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{merge_query}
\alias{merge_query}
\title{Gather and flatten querys}
\usage{
merge_query(.)
}
\arguments{
\item{.}{\code{(list)} of tsymbles}
}
\description{
Receives multiple tsymble's to be bound together and combines the query's as individual list items
}
\keyword{Internal}
