% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{msg}
\alias{msg}
\title{Create a message (\code{msg}) from an event}
\usage{
msg(event, msg = NULL, channel = NULL)
}
\arguments{
\item{event}{\code{(list)} sole argument to the websocket \code{onMessage} callback}

\item{msg}{\code{(character)} A message to be used if the event has no corresponding message}

\item{channel}{\code{(character)} a default for channel if it cannot be assumed from the event. See \verb{\link[AlpacaforR]\{event_channel\}}}
}
\description{
Create a message (\code{msg}) from an event
}
\keyword{Internal}
