% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{o_transform}
\alias{o_transform}
\title{Transform order objects}
\usage{
o_transform(.o)
}
\arguments{
\item{.o}{An order object}
}
\description{
Replaces character quantities with numeric and character dates with POSIXct
}
\keyword{internal}
