% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_cancel}
\alias{order_cancel}
\alias{cancel_orders}
\title{Cancel Order function}
\usage{
order_cancel(ticker_id = NULL, live = FALSE, v = 2)

cancel_orders(ticker_id = NULL, live = FALSE, v = 2)
}
\arguments{
\item{ticker_id}{\code{(character)} The ticker symbol(s) or order ID(s). Use \code{"cancel_all"} to cancel all open orders.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
Order \code{(tibble)} with details about canceled orders. This function places the request to cancel an order. In most cases the order is canceled instantly, but sometimes it's not. Connect to the \href{https://alpaca.markets/docs/api-documentation/api-v2/streaming/}{Alpaca Streaming API} via the websocket functions to receive order updates (See \code{\link[AlpacaforR]{ws_create}}).

Query \code{(list)} attribute accessed via \code{attr(df, "query")} on the returned \code{df} object which contains:
\itemize{
\item{\code{id}}{\code{(character)} vector of cancel order ids}
\item{\code{status}}{\code{(integer)} vector of cancel order statuses}
}
}
\description{
Cancels any open order by order_id or all open orders. Use "cancel_all" as the order_id to cancel all open orders. See \href{https://alpaca.markets/docs/api-documentation/api-v2/orders#cancel-all-orders}{Orders: DELETE} in the Alpaca API V2 documentation for details.

\code{cancel_orders} is deprecated. Use \code{\link[AlpacaforR]{order_cancel}} instead.
}
\examples{
# Cancel by the order id
order <- order_submit("AAPL", 1, "buy", "market", live = F)
order_cancel(order_id = order$id, live = F)
# Or you can use orders_get to cancel all orders for a ticker symbol:
order_submit("AAPL", 1, "buy", "market", live = F)
order_submit("AMZN", 1, "buy", "market", live = F)
canceled_orders <- order_cancel(c("AAPL", "AMZN"))
str(canceled_orders)
# Or cancel all orders:
order_submit("AAPL", 1, "buy", "market", live = F)
order_submit("AMZN", 1, "buy", "market", live = F)
order_submit("FB", 1, "buy", "market", live = F)
canceled_orders <- order_cancel("cancel_all")
str(canceled_orders)
cancel_orders()
}
\seealso{
Other Orders: 
\code{\link{order_replace}()},
\code{\link{order_submit}()},
\code{\link{orders_get}()},
\code{\link{submit_orders}()}

Other Orders: 
\code{\link{order_replace}()},
\code{\link{order_submit}()},
\code{\link{orders_get}()},
\code{\link{submit_orders}()}
}
\concept{Orders}
