% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_check}
\alias{order_check}
\title{order_check}
\usage{
order_check(
  ...,
  ovar = get0("ovar", mode = "environment", envir = rlang::caller_env())
)
}
\arguments{
\item{...}{named arguments. Will automatically get arguments from enclosing environment.}

\item{penv}{\code{environment} the parent environment, otherwise a named list of arguments from the parent environment}
}
\value{
\code{(list)} returns list with appropriate arguments, to be merged with parent environment via \code{list2env}
}
\description{
smart detect: type, order_class, extended_hours. Fix names for take_profit, stop_loss if partialled. Throw errors/warnings for specific criteria
}
\keyword{internal}
