% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_submit}
\alias{order_submit}
\title{Submit Orders}
\usage{
order_submit(
  ticker,
  qty,
  side,
  type = NULL,
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  extended_hours = FALSE,
  live = FALSE,
  v = 2
)
}
\arguments{
\item{ticker}{\code{(character)} \emph{required} The stock's symbol.}

\item{qty}{\code{(integer)} \emph{required} The amount of shares to trade.}

\item{side}{\code{(character)} \emph{required} The side of the trade. I.E "buy" or "sell"}

\item{type}{\code{(character)} The type of trade order. I.E "market","limit","stop","stop_limit", etc. This argument is assumed to be "market" unless a \code{limit_price}, \code{stop_price}, or both are set. If only \code{limit} is set, "limit" is assumed to be the type. If only \code{stop} is set, "stop" is assumed to be the type. If both are set "stop_limit" is assumed to be the type. If this argument is specified, and the requisite arguments are not set, the function will stop with error.}

\item{time_in_force}{\code{(character)} The time in force for the order. I.E "day", "gtc", "opg". In the V2 API, Immediate Or Cancel (IOC) & Fill or Kill (FOK) is added. Default is "day". Please see \href{https://alpaca.markets/docs/trading-on-alpaca/orders/#time-in-force}{Understand Orders} for more info.}

\item{extended_hours}{\code{(logical)} Default: \code{FALSE}. If \code{TRUE}, order will be eligible to execute in premarket/afterhours. Only works with type limit and time_in_force day on the V2 API.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}

\item{limit}{\code{(numeric)} If order type was a limit, then enter the limit price here.}

\item{stop}{\code{(numeric)} If order type was a stop, then enter the stop price here.}
}
\value{
Order \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/orders/#order-entity}{Order object} or Array of Order Objects with the following information:
\itemize{
\item{\code{id}}{\code{(character)} order id}
\item{\code{client_order_id}}{\code{(character)} client unique order id}
\item{\code{created_at}}{\code{(POSIXct)} nullable When the order was created}
\item{\code{updated_at}}{\code{(POSIXct)} nullable When the order was updated}
\item{\code{submitted_at}}{\code{(POSIXct)} nullable When the order was submitted}
\item{\code{filled_at}}{\code{(POSIXct)} nullable When the order was filled}
\item{\code{expired_at}}{\code{(POSIXct)} nullable When the order was expired}
\item{\code{canceled_at}}{\code{(POSIXct)} nullable When the order was canceled}
\item{\code{failed_at}}{\code{(POSIXct)} nullable When the order failed}
\item{\code{replaced_at}}{\code{(POSIXct)} nullable When the order was replaced}
\item{\code{replaced_by}}{\code{(character)} id of replacement order}
\item{\code{replaces}}{\code{(character)} id of the replaced order}
\item{\code{asset_id}}{\code{(character)} asset ID}
\item{\code{symbol}}{\code{(character)} Asset symbol}
\item{\code{exchange}}{\code{(character)} Asset exchange}
\item{\code{asset_class}}{\code{(character)} Asset class}
\item{\code{qty}}{\code{(integer)} Ordered quantity}
\item{\code{filled_qty}}{\code{(integer)} Filled quantity}
\item{\code{filled_avg_price}}{\code{(numeric)} Filled average price}
\item{\code{order_class}}{\code{(character)}}
\item{\code{order_type}}{\code{(character)}}
\item{\code{type}}{\code{(character)} Valid values: market, limit, stop, stop_limit}
\item{\code{side}}{\code{(character)} Valid values: buy, sell}
\item{\code{time_in_force}}{\code{(character)} time in force}
\item{\code{limit_price}}{\code{(numeric)} Limit price}
\item{\code{stop_price}}{\code{(numeric)} Stop price}
\item{\code{status}}{\code{(character)} Status of the order}
\item{\code{extended_hours}}{\code{(logical)} If true, eligible for execution outside regular trading hours}.
\item{\code{legs}}{\code{(character)} When querying non-simple order_class orders in a nested style, an array of Order entities associated with this order. Otherwise, null.
}
}
\description{
Places a new order of the specified stock, quantity, buy / sell, type of order, time in force, and limit / stop prices if selected. See \href{https://alpaca.markets/docs/api-documentation/api-v2/orders#request-a-new-order}{Orders: POST}.
}
\examples{
# For market order (`type` is assumed to be "market" when `stop_price` and `limit_price` are not specified):
order_submit(ticker = "AAPL", qty = 100, side = "buy")
# Or you can submit a limit order (`type` is assumed to be "limit" when only `limit_price` is set):
order_submit(ticker = "AAPL", qty = 100, side = "sell", lim = 120)
}
\seealso{
Other Orders: 
\code{\link{order_cancel}()},
\code{\link{order_replace}()},
\code{\link{orders_get}()},
\code{\link{submit_orders}()}
}
\concept{Orders}
