% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_symbol_id}
\alias{order_symbol_id}
\title{retrieve the order id if an order object is supplied}
\usage{
order_symbol_id(
  symbol_id,
  ...,
  ovar = get0("ovar", mode = "environment", envir = rlang::caller_env())
)
}
\arguments{
\item{symbol_id}{\code{(character)}  The stock symbol (\emph{Required} when \code{action = "submit"}) or Order object (single row tibble) (\emph{Required} when \code{action = "cancel"/"replace"}).
To expedite the setting of stops and limits for open positions, an Order ID from a \code{'buy'} order can be provided when \code{action = "submit"} to place a \code{'sell'} order with the following parameters such that they do not need to be set manually:
\itemize{
\item{\code{side = 'sell'}}
\item{If \code{qty} is not provided, it will be populated from the buy order}
\item{\code{symbol_id} will be set to the symbol from the buy order.}
\item{If \code{client_order_id = TRUE}, the \code{client_order_id} will be set to the buy Order ID provided, effectively linking the orders for your records.}
\item{All other parameters can be specified as usual.}
}}
}
\description{
Retrieves the order_id from an order object if provided as \code{symbol_id}, or provides informative error if input order failed.
}
\keyword{internal}
