% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_transform}
\alias{order_transform}
\title{Transform order responses}
\usage{
order_transform(o)
}
\arguments{
\item{orders}{A dataframe returned from any orders_* endpoint}
}
\value{
\code{(tibble)}  with respective R compliant objects (numeric, POSIXct/Datetime, character)
}
\description{
Parses order type responses and replaces plain text quantities and dates with respective R objects
}
\keyword{internal}
