% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{submit_orders}
\alias{submit_orders}
\title{submit_orders}
\usage{
submit_orders(
  ticker_id,
  action = "submit",
  qty = NULL,
  side = NULL,
  type = NULL,
  time_in_force = "day",
  limit = NULL,
  stop = NULL,
  extended_hours = NULL,
  client_order_id = NULL,
  order_class = NULL,
  take_profit = NULL,
  stop_loss = NULL,
  live = FALSE,
  v = 2
)
}
\description{
\code{submit_orders} is deprecated. Use \code{\link[AlpacaforR]{order_submit}} instead.
}
\seealso{
Other Orders: 
\code{\link{order_cancel}()},
\code{\link{order_replace}()},
\code{\link{order_submit}()},
\code{\link{orders}()}
}
\concept{Orders}
