% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{orders_transform}
\alias{orders_transform}
\title{Transform order objects

Replaces plain text quantities and dates with respective R objects}
\usage{
orders_transform(orders)
}
\arguments{
\item{orders}{A dataframe returned from any orders_* endpoint}
}
\value{
\code{(tibble)}  with respective R compliant objects (numeric, POSIXct/Datetime, character)
}
\description{
Transform order objects

Replaces plain text quantities and dates with respective R objects
}
\keyword{internal}
