% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{poly_transform}
\alias{poly_transform}
\title{transform polygon.io reference endpoints}
\usage{
poly_transform(resp, ep)
}
\arguments{
\item{resp}{The response object from httr}

\item{ep}{The endpoint}
}
\value{
\code{list/data.frame/tibble} Either a list or tibble depending on the endpoint
}
\description{
coerce data objects returned from the various polygon.io endpoints to R compliant objects
}
\keyword{internal}
