% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{port_transform}
\alias{port_transform}
\title{transform portfolio history}
\usage{
port_transform(resp)
}
\arguments{
\item{resp}{\code{(response)} The response object}
}
\value{
A \href{https://alpaca.markets/docs/api-documentation/api-v2/portfolio-history/#portfoliohistory-entity}{PortfolioHistory} \code{(tibble)} Object with R amenable data types. \emph{Note} that base value & timeframe can be accessed by calling \code{attr(out, "info")} on the returned tibble.
\itemize{
\item{\code{timestamp}}{ \code{(Date/POSIXlt/Datetime(POSIXct)/character)} time of each data element, with the first element being the beginning of the time window.}
\item{\code{equity}}{ \code{(numeric)} equity value of the account in dollar amount as of the end of each time window}
\item{\code{profit_loss}}{ \code{(numeric)} profit/loss in dollar from the base value}
\item{\code{profit_loss_pct}}{ \code{(numeric)} profit/loss in percentage from the base value}
\item{\code{base_value}}{ \code{(numeric)} basis in dollar of the profit loss calculation}
\item{\code{timeframe}}{ \code{(character)} time window size of each data element}
}
}
\description{
Transforms a \href{https://alpaca.markets/docs/api-documentation/api-v2/portfolio-history/#portfoliohistory-entity}{PortfolioHistory} response object returned from \code{account_portfolio}.
}
\keyword{internal}
