% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pos_transform}
\alias{pos_transform}
\title{transform positions objects

Cleans arrays of position objects, and position objects}
\usage{
pos_transform(pos)
}
\description{
transform positions objects

Cleans arrays of position objects, and position objects
}
\keyword{internal}
