% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positions.R
\name{positions}
\alias{positions}
\title{Get Positions}
\usage{
positions(symbols = NULL, action = "get", live = get_live())
}
\arguments{
\item{symbols}{\code{(character)} Symbol(s) for which Position information will be retrieved.}

\item{action}{\code{(character)}
\itemize{
\item{\code{"get"/"g"}}{ \code{GET} all positions or those specified by \code{symbols}}
\item{\code{"close","c"}}{ Close positions specified by \code{symbols}}
\item{\code{"close_all"}}{ Close all positions}
}}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
Position \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/#position-entity}{Position object} or array of Position objects of length 0 if no positions, otherwise with length of the number of positions, each with the following attributes:
\itemize{
\item{\code{asset_id}}{\code{(character)} Asset ID.}
\item{\code{symbol}}{\code{(character)} Symbol of the asset.}
\item{\code{exchange}}{\code{(character)} Exchange name of the asset.}
\item{\code{asset_class}}{\code{(character)} Asset class name.}
\item{\code{qty}}{\code{(integer)} The number of shares.}
\item{\code{avg_entry_price}}{\code{(numeric)} Average entry price of the position.}
\item{\code{side}}{\code{(character)} long/short exposure.}
\item{\code{market_value}}{\code{(numeric)} Total dollar amount of the position.}
\item{\code{cost_basis}}{\code{(numeric)} Total cost basis in dollar.}
\item{\code{unrealized_pl}}{\code{(numeric)} Unrealized profit/loss in dollar.}
\item{\code{unrealized_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{unrealized_intraday_pl}}{\code{(numeric)} Unrealized profit/loss in dollar for the day.}
\item{\code{unrealized_intraday_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{current_price}}{\code{(numeric)} Current asset price per share.}
\item{\code{lastday_price}}{\code{(numeric)} Last day's asset price per share.}
\item{\code{change_today}}{\code{(numeric)} Percent change from last day price (by a factor of 1).}
}
}
\description{
The positions API provides information about an account's current open positions. The response will include information such as cost basis, shares traded, and market value, which will be updated live as price information is updated. Character values are returned as a string while integer values are returned as numeric. See \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/}{Positions} for details.
}
\details{
\code{position} is vectorized and thus multiple arguments may be provided to \code{symbols} and the function will perform the \code{action} specified for each. \code{action} is not vectorized, and only one action may be performed for a set of symbols at a given time.
}
\examples{
positions() # all paper positions
positions(live = TRUE) # all live positions
#cancel all paper positions
positions(a = "close_all")
\dontrun{
# details for a specific position (errors if not present)
positions("GOOG")
}
}
\concept{Positions}
