% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{replace_order}
\alias{replace_order}
\title{Replace Order function}
\usage{
replace_order(
  ticker_id,
  qty = NULL,
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  live = FALSE
)
}
\arguments{
\item{ticker_id}{The ticker symbol or the order id.}

\item{qty}{The amount of shares to replace.}

\item{time_in_force}{The type of time order. I.E "day", "gtc", "opg". Default is "day".}

\item{limit_price}{If order type was a limit, then enter the limit price here.}

\item{stop_price}{If order tyope was a stop, then enter the stop price here.}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\description{
Cancels any open order by either ticker or order id. If multiple open orders exist for one ticker, then the default is to cancel the most recent order.
}
\examples{
replace_order(ticker_id = "AAPL", qty = <new qty amount>, time_in_force = <new time in force>, limit_price=<new limit price, if applicable>, stop_price=<new stop price, if applicable>, live = FALSE)
Or you can instead cancel by the order_id:
orders <- get_orders(status="open", silent = TRUE)
replace_order(ticker_id = orders$id[1], qty = <new qty amount>, time_in_force = <new time in force>, limit_price=<new limit price, if applicable>, stop_price=<new stop price, if applicable>, live = FALSE)
}
