% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{set_config}
\alias{set_config}
\title{Send Account Configurations function}
\usage{
set_config(
  live = FALSE,
  dtbp_check = "entry",
  no_shorting = FALSE,
  suspend_trade = FALSE,
  trade_confirm_email = "all"
)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
The response from the server after setting the configuration.
}
\description{
Send Account Configurations function
}
\examples{
set_config(live = FALSE, dtbp_check = "entry", no_shorting = FALSE, suspend_trade = FALSE, trade_confirm_email = "all")
# Which is similar to:
set_config(dtbp_check = "entry", no_shorting = FALSE, suspend_trade = FALSE, trade_confirm_email = "all")
# For access to live accounts, you must submit as live = TRUE
set_config(live = TRUE, dtbp_check = "entry", no_shorting = FALSE, suspend_trade = FALSE, trade_confirm_email = "all")
}
