% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{socket_detect}
\alias{socket_detect}
\title{Determine the socket from the channel}
\usage{
socket_detect(channel)
}
\arguments{
\item{channel}{\code{(character)} Name of the channel}
}
\value{
\code{(character)} socket name based on channel
}
\description{
Determine the socket from the channel
}
