% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{tf_num}
\alias{tf_num}
\title{convert timeframe to numeric \code{.tf_num}}
\usage{
tf_num(timeframe, ..., cenv = rlang::caller_env())
}
\arguments{
\item{timeframe}{\code{(character)} For the \code{'v1'} API, one of
\itemize{
\item{\code{'m'}/\code{'min'}/\code{'minute'}}{ (\code{multiplier} can be \code{1}/\code{5}/\code{15})}
\\item{\code{'d'}/\code{'day'}}{ (\code{multiplier} will be \code{1})}
}
Not case-sensitive.
For the \code{'v2'} API, \code{multiplier} can be any \code{integer} for any one of the following \code{timeframe}'s:
\itemize{
\item{\code{'m'}/\code{'min'}/\code{'minute'}}
\item{\code{'h'}/\code{'hour'}}
\item{\code{'d'}/\code{'day'}}
\item{\code{'w'}/\code{'week'}}
\item{\code{'M'}/\code{'mo'}/\code{'month'}}{ (\emph{Note} capitalized M for month)}
\item{\code{'q'}/\code{'quarter'}}
\item{\code{'y'}/\code{'year'}}
}
Case-sensitive}

\item{...}{additional arguments}

\item{cenv}{Calling environment in which to assign variables}
}
\value{
.tf_num & timeframe in calling environment.
}
\description{
If \code{.tf_num} is not found in environment, create it from \code{timeframe}. Throw errors if arguments are unacceptable
}
\keyword{internal}
