% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{time_index}
\alias{time_index}
\title{time_index}
\usage{
time_index(x, type = "character")
}
\arguments{
\item{x}{\code{(tsibble/list with tibble)}}

\item{type}{\itemize{
\item{\code{"character"/"c"}}{ (Default) \code{(character)} vector of the index}
\item{\code{"value"/"v"}}{ \code{(Date/Datetime)} vector of values in the time index}
\item{\code{"language"/"symbol"/"s"/"l"}}{ \code{(symbol)} for the time index}
}}
}
\value{
\code{(character/numeric)} Either the name of the index, or the index itself based on \code{name} argument
}
\description{
Get the time index for a data.frame or xts
}
