% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{time_interval}
\alias{time_interval}
\title{time_interval}
\usage{
time_interval(x)
}
\arguments{
\item{x}{\code{(data.frame/tsibble)} input data}
}
\value{
\code{(list)} with the following:
\itemize{
\item{multiplier}{ \code{(numeric)} of the multiple of the interval period}
\item{timeframe}{ \code{(character)} describing the interval}
}
}
\description{
Retrieve the time interval. If a \code{tsibble}, values are retrieved with \code{tsibble::interval}. If another object, the time index is detected and the Mode of the intervals is used.
}
