% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{toNum}
\alias{toNum}
\title{Convert money strings to numeric

remove $ from cash quantity strings and convert to numeric}
\usage{
toNum(x)
}
\description{
Convert money strings to numeric

remove $ from cash quantity strings and convert to numeric
}
\keyword{internal}
