% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{transform_bars}
\alias{transform_bars}
\title{transform_bars}
\usage{
transform_bars(bars)
}
\description{
Internal function for transforming data from Alpaca API to a human-readable TTR/quantmod compatible format
}
\keyword{internal}
