% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{try_date}
\alias{try_date}
\title{Reliable Date/Datetime conversion for most time representations from the \href{https://polygon.io/}{Polygon.io} & \href{https://alpaca.markets/}{Alpaca.markets} APIs}
\usage{
try_date(.x, timeframe = "day", tz = NULL)
}
\arguments{
\item{.x}{\code{(numeric/character/Date/Datetime)} input vector}

\item{timeframe}{\code{(character)} The timeframe of the \code{.x} vector. See \link[AlpacaforR]{market_data}.}

\item{tz}{\code{(character)} Timezone. See \link[base]{OlsonNames} for valid strings. \strong{Default NULL}}
}
\description{
Attempts to coerce input to a valid \code{\link[lubridate]{Date} or \link[lubridate]{POSIXct}} object.
}
