% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{update_watchlist}
\alias{update_watchlist}
\title{Update Watchlist function}
\usage{
update_watchlist(
  watchlist_id = NULL,
  name = NULL,
  tickers = c(NULL),
  live = FALSE
)
}
\arguments{
\item{watchlist_id}{Watchlist ID <string>}

\item{name}{arbitrary name string, up to 64 characters. <string>}

\item{tickers}{Set of symbols. <vector of strings>}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
A watchlist object with updated content.
}
\description{
Update the name and/or content of watchlist. For symbols, You must provide current tickers, and new ticker to add.
}
\examples{
update_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", name = "test_watch", tickers = c("AAPL","WMT"), live=FALSE)
}
