% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{url_to_path}
\alias{url_to_path}
\title{Convert url to valid file path}
\usage{
url_to_path(url)
}
\arguments{
\item{url}{\code{(character)} url to convert to file path}
}
\description{
Convert url to valid file path
}
