% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R
\name{watchlist_update}
\alias{watchlist_update}
\title{Update Watchlist function}
\usage{
watchlist_update(
  watchlist_id = NULL,
  name = NULL,
  tickers = NULL,
  action = "add",
  live = FALSE,
  v = 2
)
}
\arguments{
\item{watchlist_id}{\code{(character)} Name or ID of watchlist to get or delete,  or if \code{tickers} is non NULL, to add tickers to, delete tickers from, or replace tickers of. If \code{NULL} (Default) the array of all \code{Watchlist} will be returned. If \code{NULL} and \code{name} is specified, \code{action} is assumed to be \code{"create"} and a \code{Watchlist} named \code{name} will be created with any \code{tickers} specified. All other actions to be performed on the provided \code{Watchlist} name or ID must be explicitly specified via \code{action}. See other parameters for details.
\itemize{
\item{\code{NULL}}{ If \code{watchlist_id = NULL} & \code{action = NULL}, and no other arguments specified. An array of all \code{Watchlist} will be returned. If \code{name} is specified, \code{action} is assumed to be \code{create}.}
\item{\code{"create"}}{ A \code{Watchlist} by this name will be created, unless \code{watchlist_id} is specified, in which case the specified \code{Watchlist} will be renamed preserving existing tickers. To replace tickers, \code{action} must explicitly be \code{"replace"}.}
\item{\code{"add"}}{If \code{tickers} are specified they will be added to the \code{Watchlist}.}
\item{\code{"replace"}}{If \code{tickers} are specified they will replace existing tickers or if \code{tickers = NULL}, all tickers will be removed.}
\item{\code{"delete"}}{ If no tickers are specified, the \code{Watchlist} specified by \code{watchlist_id} will be deleted. If \code{tickers} are specified, the \code{Watchlist} specified by \code{watchlist_id} will have the specified tickers deleted from it.}
}}

\item{name}{\code{(character)} arbitrary replacement name, up to 64 characters.}

\item{tickers}{\code{(character)} Set of symbols to add, replace or delete those in the watchlist according to the \code{action} argument. See Details for specifics}

\item{action}{\code{(character)} Action in reference to symbols specified by \code{tickers}. Whether to add ("add"/"a") them to the existing,  replace ("replace","r") all existing assets in the watchlist with them or delete ("delete"/"d") them from the existing symbols in a watchlist. Case-insensitive. Default \code{"add"}. See Details for specifics.}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
\code{Watchlist} \code{(tibble)} If \code{action = "create"/NULL} a \href{https://alpaca.markets/docs/api-documentation/api-v2/watchlist/#watchlist-entity}{Watchlist} Object or array of \code{Watchlist} objects. For all others \code{action}s, an \href{https://alpaca.markets/docs/api-documentation/api-v2/assets/#asset-entity}{Asset} Object or array of Asset objects with the \code{Watchlist} stored as an attribute retrieved with \code{attr(Asset, "info")}. A \code{Watchlist} object has the following information:
\itemize{
\item{\code{name}}{\code{(character)} The name of the watchlist.}
\item{\code{updated_at}}{\code{(POSIXct)} The timestamp of the last update.}
\item{\code{id}}{\code{(character)} Watchlist ID}
\item{\code{account_id}}{\code{(character)} The account ID associated with the watchlist}
\item{\code{created_at}}{\code{(POSIXct)} The timestamp of the watchlist creation.}
}
}
\description{
Update the name and/or content of watchlist. For symbols, You must provide current tickers, and new ticker to add.
}
\details{
The default \code{action} is \code{"add"} which will add any \code{tickers} specified to the watchlist indicated with \code{watchlist_id}. When a new watchlist \code{name} is specified in addition to \code{action = "add"} (Default), existing symbols in the watchlist will be preserved if no \code{tickers} are specified. If \code{tickers} are provided they will be added after the watchlist is renamed. If \code{action = "replace"}, symbols in the specified watchlist will be replaced by those provided to the \code{tickers} argument. If \code{action = "delete"}, the symbols specified by \code{tickers} will be removed from the watchlist. If a new \code{name} is provided concurrently, the \code{name} will first be changed, then the \code{tickers} will be removed.
}
\examples{
watchlist("test", tickers = c("AAPL", "WMT"))
# Update the watchlist name to test2 and replace the symbols with "FB", "AAPL", "AMZN", "NFLX", "GOOG"
(wl <- watchlist_update("test", name = "test2", tickers = c("FB", "AAPL", "AMZN", "NFLX", "GOOG"), action = "r"))
# Walmart needn't be in there, remove it and rename the watchlist appropriately
(wl <- watchlist_update("test2", name = "FAANG", ticker = "WMT", action = "d"))
# The next iteration of "FAANG", "FABANGG"
(wl <- watchlist_update("FAANG", "FABANGG", ticker = c("BYND", "GOOGL"), action = "a"))
# Back to the original FAANG
(wl <- watchlist_update("FABANGG", tickers = c("BYND", "GOOGL"), action = "d"))
# rename it appropriately
(wl <- watchlist_update("FABANGG", "FAANG"))
attr(wl, "info")
# remove it
watchlist("FAANG", a = "d")
}
\seealso{
Other Watchlist: 
\code{\link{add_to_watchlist}()},
\code{\link{delete_from_watchlist}()},
\code{\link{delete_watchlist}()},
\code{\link{watchlist}()}
}
\concept{Watchlist}
\keyword{internal}
