% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{wl_nm2id}
\alias{wl_nm2id}
\title{Look up watchlist id from name}
\usage{
wl_nm2id(nm, ..., e = environment())
}
\arguments{
\item{nm}{\code{(character)} \emph{required} the name of the watchlist}
}
\value{
id \code{(character)} the id of the watchlist OR, if no id, the array of watchlists
}
\description{
fetch the watchlist id corresponding to a watchlist name
}
\keyword{internal}
