% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R
\name{wl_transform}
\alias{wl_transform}
\title{Transform watchlist objects}
\usage{
wl_transform(resp, action)
}
\arguments{
\item{resp}{The response}
}
\value{
\code{(tibble)} with respective R compliant objects (POSIXct)
}
\description{
Takes the watchlist endpoints server response and transforms to R amenable output
}
\keyword{internal}
