% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Websockets.R
\name{write_prep}
\alias{write_prep}
\title{Prepare to record websocket data}
\usage{
write_prep(write_dir, url, channel, overwrite, private)
}
\arguments{
\item{write_dir}{\code{(character/logical)} \strong{Default: "AlpacaStreams"} top-level directory in which to store log files.}

\item{channel}{\code{(character)} The name of the channel (will be used as the filename)}

\item{overwrite}{\code{(logical)} \strong{Default: \code{FALSE}} Indicator for whether to overwrite previous log files by making new files (\code{TRUE}) or to leave old files (\code{FALSE})}

\item{private}{\code{(environment)} The \verb{*Socket} private environment}
}
\description{
Create folders & files for recording based on inputs
}
\keyword{Internal}
