% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca_websockets.R
\name{ws_auth}
\alias{ws_auth}
\title{ws_auth: A function to authenticate a websocket with the Alpaca API}
\usage{
ws_auth(ws)
}
\arguments{
\item{ws}{\code{(environment)} The websocket environment created with \link[AlpacaforR]{`ws_create`}.}
}
\value{
Returns nothing. Provided the websocket is properly connected by calling \code{\link[AlpacaforR]{ws_create}}, the console will print the server response message outlined in the \href{https://alpaca.markets/docs/api-documentation/api-v2/streaming/#authentication}{websocket documentation}.
}
\description{
Given the `websocket environment` the function will authenticate using the environment variables `APCA-LIVE-API-KEY-ID` and `APCA-LIVE-API-SECRET-KEY`. If these are unset, please see the \href{https://github.com/jagg19/AlpacaforR}{package documentation} for instruction on setting them.
}
