% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca_websockets.R
\name{ws_listen}
\alias{ws_listen}
\title{ws_listen: A function to set listening streams with the Alpaca websocket API}
\usage{
ws_listen(ws, trade = T, account = T)
}
\arguments{
\item{ws}{\code{(environment)} The websocket environment created with \link[AlpacaforR]{`ws_create`}.}

\item{trade}{\code{logical} indicating whether to connect to the trade stream. Defaults to TRUE.}

\item{account}{\code{logical} indicating whether to connect to the account stream. Defaults to TRUE.}
}
\value{
Returns nothing. Provided the websocket is properly connected, the console will print the server response message outlined in the \href{https://alpaca.markets/docs/api-documentation/api-v2/streaming/#order-updates}{websocket documentation}.
}
\description{
Given the `websocket environment` the function will listen to the specified websocket streams.
}
