% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ws_log}
\alias{ws_log}
\title{ws_log}
\usage{
ws_log(out, ..., .o = NULL, msg = NULL, penv = rlang::caller_env())
}
\arguments{
\item{out}{\code{(list)} The ws_create out object}

\item{.o}{\code{(list)} The raw message content from the Websocket}

\item{log_bars}{\code{(logical)} The flag as to whether to log bars on the drive as CSV or not}
}
\value{
bars \code{(list)} object in the out$env environment in the object returned from \code{ws_create} with the previously transmitted data as a \code{tibble} for each polygon subscription channel, each named according to the channel from which it came. Additionally, a CSV named by the Subscription channel if \code{logbars = T} in the local or specified directory with the same data.
}
\description{
Performs logging of streaming bars data based on input options to ws_create
}
\details{
The rows of the each of the bars objects are halved if it's size reaches .33 of the memory allocated to R. Prevents memory overflow and potential freezing.
}
\keyword{internal}
