% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ws_msg}
\alias{ws_msg}
\title{Update Websocket message objects}
\usage{
ws_msg(out, .o = NULL, msg, toConsole = T)
}
\description{
Update Websocket message objects
}
\keyword{Used}
\keyword{and}
\keyword{in}
\keyword{instantiate}
\keyword{internal}
\keyword{message}
\keyword{objects}
\keyword{to}
\keyword{update}
\keyword{websocket}
\keyword{ws_create}
