% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Account.R
\name{account_config}
\alias{account_config}
\alias{get_config}
\alias{set_config}
\title{Get Account Configurations}
\usage{
account_config(
  dtbp_check = NULL,
  no_shorting = NULL,
  pdt_check = NULL,
  suspend_trade = NULL,
  trade_confirm_email = NULL,
  live = FALSE,
  v = 2
)

get_config(live = FALSE)

set_config(
  live = FALSE,
  dtbp_check = "entry",
  no_shorting = FALSE,
  suspend_trade = FALSE,
  trade_confirm_email = "all"
)
}
\arguments{
\item{dtbp_check}{\code{(character)} both, entry, or exit. Controls Day Trading Margin Call (DTMC) checks. Default NULL (no change). Set to "default" to set all configuration options back to defaults. See Value section for defaults.}

\item{no_shorting}{\code{(logical)} If true, account becomes long-only mode. Default NULL (no change).}

\item{pdt_check}{\code{(character)} Not documented.}

\item{suspend_trade}{\code{(logical)} If true, new orders are blocked. Default NULL (no change).}

\item{trade_confirm_email}{\code{(character)} all or none. If none, emails for order fills are not sent. Default NULL (no change).}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
AccountConfigurations \code{(list)} \href{https://alpaca.markets/docs/api-documentation/api-v2/account-configuration/#accountconfigurations-entity}{AccountConfigurations}  object with the following parameters (defaults shown):
\itemize{
\item{\code{dtbp_check = "entry"}}
\item{\code{no_shorting = FALSE}}
\item{\code{pdt_check = "entry"}}
\item{\code{suspend_trade = FALSE}}
\item{\code{trade_confirm_email = "all"}}
}
}
\description{
The account configuration API serves an important role in setting the way you want. See \href{https://alpaca.markets/docs/api-documentation/api-v2/account-configuration/}{Account Configurations} for details.

\code{get_config} is deprecated, please use \code{\link[AlpacaforR]{account_config}} instead.

\code{set_config} is deprecated. Please use \code{\link[AlpacaforR]{account_config}} instead.
See \href{https://alpaca.markets/docs/api-documentation/api-v2/account-configuration/}{Account Configurations} for details.
}
\examples{
account_config(live = FALSE)
# Which is similar to:
account_config()
# For access to live accounts, you must submit as live = TRUE
account_config(live = TRUE)
account_config(dtbp_check = "both", no_shorting = T)
accont_config("default")
get_config()
}
\seealso{
Other Account: 
\code{\link{account_activities}()},
\code{\link{account}()}
}
\concept{Account}
