% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{add_row.tsymble_ts}
\alias{add_row.tsymble_ts}
\title{Add a row to a tsymble}
\usage{
add_row.tsymble_ts(.data, ..., .before = NULL, .after = NULL)
}
\arguments{
\item{.data}{Data frame to append to.}

\item{...}{
  Arguments passed on to \code{\link[tibble:add_row]{tibble::add_row}}
  \describe{
    \item{\code{.before,.after}}{One-based row index where to add the new rows,
default: after last row.}
  }}

\item{.before, .after}{One-based row index where to add the new rows,
default: after last row.}
}
\description{
\lifecycle{experimental}
Adds a row to a tsymble
}
\keyword{internal}
