% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R, R/alpaca4R.R
\name{add_to_watchlist}
\alias{add_to_watchlist}
\alias{watchlist_update}
\title{Add to Watchlist function (Deprecated)}
\usage{
add_to_watchlist(watchlist_id = NULL, tickers = NULL, live = FALSE)

add_to_watchlist(watchlist_id = NULL, tickers = NULL, live = FALSE)
}
\arguments{
\item{watchlist_id}{Watchlist ID <string>}

\item{tickers}{Set of symbols. <vector of strings REQUIRED>}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
A watchlist object with updated content.

A watchlist object with updated content.
}
\description{
\code{add_to_watchlist} is deprecated. Use \code{\link[AlpacaforR]{watchlist}} with \code{action = "add"} instead. Append an asset for the symbol to the end of watchlist asset list. You must provide current tickers, and new ticker to add.

Append an asset for the symbol to the end of watchlist asset list. You must provide current tickers, and new ticker to add.
}
\examples{
add_to_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", tickers = c("AAPL","WMT"), live = FALSE)
add_to_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", tickers = c("AAPL","WMT"), live = FALSE)
}
\seealso{
Other Watchlist: 
\code{\link{delete_from_watchlist}()},
\code{\link{delete_watchlist}()},
\code{\link{watchlist_update}()},
\code{\link{watchlist}()}
}
\concept{Watchlist}
