% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_complete}
\alias{bars_complete}
\title{Query the API as many times as necessary to fill a market_data query}
\usage{
bars_complete(bars, .missing, ...)
}
\arguments{
\item{\code{bars}}{Returned by \code{bars_get}}

\item{\code{.missing}}{Returned by \code{bars_missing}}
}
\description{
This function uses the output from bars_missing and fills the missing data. It's arguments are represented in the first object \code{.vn}.
}
\details{
The arguments for all bars_\* functions  will be called from the environment from which the function is called if not specified explicitly. If params need to be specified explicitly, \strong{all params must be named}.
}
\keyword{internal}
