% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_get}
\alias{bars_get}
\title{Retrieve market data from the respective API}
\usage{
bars_get(url, ...)
}
\arguments{
\item{url}{\code{(character)} the url rendered by \code{\link[AlpacaforR]{bars_url}}}
}
\value{
\code{(list)} See \code{\link[AlpacaforR]{market_data}} as the output is the same.
}
\description{
Internal function for retrieving data from Alpaca API and outputting data.frame with query attribute. See notes on specifying arguments in the \emph{Details} section for \code{\link[AlpacaforR]{bars_complete}}.
}
\seealso{
Other bars: 
\code{\link{bars_bounds}()},
\code{\link{bars_complete}()},
\code{\link{bars_expected}()},
\code{\link{bars_missing}()},
\code{\link{bars_transform}()},
\code{\link{bars_url}()}
}
\concept{bars}
\keyword{internal}
