% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_missing}
\alias{bars_missing}
\title{Determine missing values in bars objects}
\usage{
bars_missing(bars, ..., .tf_reduce = F)
}
\arguments{
\item{bars}{Output from \code{\link[AlpacaforR]{bars_get}}}

\item{v}{See \code{\link[AlpacaforR]{market_data}}}

\item{.bounds}{Output from \code{\link[AlpacaforR]{bars_bounds}}}
}
\value{
\code{(list)} If there are no missing values, NULL is returned. Otherwise, a list with two items corresponding to the data for each symbol returned:
\itemize{
\item{\code{missing} vector of missing expected Date/Datetimes}
\item{\code{gaps}}{ \code{(data.frame)} with three columns:
\itemize{
\item{\code{gap_times}}{ a vector of Date/Datetimes with the gaps in the data that can be used for subsequent API query bounds.}
\item{\code{gap_inds}}{ the indices of the gaps in the data (if the missing data is prior to or after the actual data, this will be NA).}
\item{\code{f}}{ a factor to \code{split} the gaps into query boundaries}
\item{\code{position}}{ a factor indicating whether the row corresponds to  \code{leading} missing data, or the \code{begin}ing or \code{end} of a gap in the data.}
}
}
}
}
\description{
Internal function for comparing actual data to expected data and returning missing data. The function is intended for use in an environment where its parameters already exist.  See notes on specifying arguments in the \emph{Details} section for \code{\link[AlpacaforR]{bars_complete}}
}
\seealso{
Other bars: 
\code{\link{bars_bounds}()},
\code{\link{bars_complete}()},
\code{\link{bars_expected}()},
\code{\link{bars_get}()},
\code{\link{bars_transform}()},
\code{\link{bars_url}()}
}
\concept{bars}
\keyword{internal}
