% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Market_Data.R
\name{bars_transform}
\alias{bars_transform}
\title{Transform bars objects}
\usage{
bars_transform(
  agg_quote,
  ...,
  evar = get0("evar", mode = "environment", envir = rlang::caller_env())
)
}
\description{
Internal function for transforming data from Alpaca API to a human-readable TTR/quantmod compatible format
}
\keyword{internal}
