% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{bars_transform}
\alias{bars_transform}
\title{Transform bars objects}
\usage{
bars_transform(bars)
}
\description{
Internal function for transforming data from Alpaca API to a human-readable TTR/quantmod compatible format
}
\keyword{internal}
