% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsymble.R
\name{bind_rows}
\alias{bind_rows}
\title{bind_rows}
\usage{
bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[dplyr:bind]{dplyr::bind_rows}}
  \describe{
    \item{\code{}}{}
  }}

\item{.id}{Data frame identifier.

When \code{.id} is supplied, a new column of identifiers is
created to link each row to its original data frame. The labels
are taken from the named arguments to \code{bind_rows()}. When a
list of data frames is supplied, the labels are taken from the
names of the list. If no names are found a numeric sequence is
used instead.}
}
\value{
\code{bind_rows()} and \code{bind_cols()} return the same type as
the first input, either a data frame, \code{tbl_df}, or \code{grouped_df}.
}
\description{
Replace \code{dplyr::bind_rows} for use with \code{tsymble_ts} objects.
}
