% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Order function}
\usage{
cancel_order(url, headers, ticker)
}
\arguments{
\item{url}{URL to your user live or paper API.}

\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}

\item{ticker}{The stock's symbol as a string.}
}
\description{
Attempts to cancel an open order. If the order is no longer cancelable (example: status=order_filled), the server will respond with status 422, and reject the request.
}
\examples{
cancel_order(url = url, headers = headers, ticker = "AMZN")
}
