% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Order function}
\usage{
cancel_order(ticker_id = NULL, all = FALSE, live = FALSE, version = "v2")
}
\arguments{
\item{ticker_id}{The ticker symbol or the order id. If all = TRUE, no ticker_id is needed because ALL orders will be canceled.}

\item{all}{Default to False. If true, all open orders are cancelled. Only available in the V2 API.}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{version}{Use the deprecated V1 API or the newer V2 API.}
}
\description{
Cancels any open order by either ticker or order id. If multiple open orders exist for one ticker, then the default is to cancel the most recent order. As of the V2 API update, an "all" arguement is added to cancel all open orders.
}
\examples{
\dontrun{
# \emph{Note} At least one order for AAPL must be placed prior to running this example for the example to work.
cancel_order(ticker_id = "AAPL", version = "v2")
cancel_order(ticker_id = "aapl", version = "v2")
# Or you can instead cancel by the order_id:
orders <- get_orders(status="open", silent = TRUE, version = "v2")
cancel_order(ticker_id = orders$id[1])
}
}
