% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{cancel_order}
\alias{cancel_order}
\title{Cancel Order function}
\usage{
cancel_order(live = FALSE, ticker, order_id = NULL)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was specified.}

\item{ticker}{The stock's symbol as a string.}

\item{order_id}{You can specify which order to cancel by entering order_id. Defaults to most recent open order.}
}
\description{
Attempts to cancel an open order. If the order is no longer cancelable (example: status=order_filled), the server will respond with status 422, and reject the request.
}
\examples{
cancel_order(ticker = "AAPL")
Or you can instead cancel by the order_id:
cancel_order(order_id = VALUE)
}
