% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positions.R, R/alpaca4R.R
\name{close_position}
\alias{close_position}
\title{Close Positions by ticker, or close all (Deprecated)}
\usage{
close_position(ticker = NULL, live = FALSE)

close_position(ticker = NULL, live = FALSE)
}
\arguments{
\item{ticker}{\code{(character)} \emph{required} the symbol of the ticker for which positions will be closed. Use "close_all" to close all positions.}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
"asset_id"  Asset ID.

"symbol"  Symbol of the asset.

"exchange"  Exchange name of the asset.

"asset_class"  Asset class name.

"avg_entry_price"  Average entry price of the position.

"qty" The number of shares.

"side" long/short exposure.

"market_value"  Total dollar amount of the position.

"cost_basis"  Total cost basis in dollar.

"unrealized_pl"  Unrealized profit/loss in dollar.

"unrealized_plpc"  Unrealized profit/loss percent (by a factor of 1).

"unrealized_intraday_pl"  Unrealized profit/loss in dollar for the day.

"unrealized_intraday_plpc"  Unrealized profit/loss percent (by a factor of 1).

"current_price"  Current asset price per share.

"lastday_price"  Last day’s asset price per share.

"change_today"  Percent change from last day price (by a factor of 1).
}
\description{
\code{close_position} is deprecated. Use \code{\link[AlpacaforR]{positions}} with \code{action = "close"} instead. The close_positions function liquidates all positions or those specified by \code{ticker}. Character values are returned as a string while integer values are returned as numeric. See \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/}{Positions} for details.

The close all positions API liquidates all currently open long and short positions. Character values are returned as a string while integer values are returned as numeric.
}
\examples{
close_position(ticker = "AAPL", live = FALSE)
close_position(ticker = "AAPL")
close_position(ticker = "AAPL", live = TRUE)
close_position(ticker = "AAPL", live = FALSE)
close_position(ticker = "AAPL")
close_position(ticker = "AAPL", live = TRUE)
}
\seealso{
Other Positions: 
\code{\link{positions}()}
}
\concept{Positions}
