% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R
\name{delete_watchlist}
\alias{delete_watchlist}
\title{Delete Watchlist function (Deprecated)}
\usage{
delete_watchlist(watchlist_id = NULL, live = FALSE)
}
\arguments{
\item{watchlist_id}{Watchlist ID}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\description{
Deprecated, use \code{\link[AlpacaforR]{watchlist}} instead. Delete a watchlist. This is a permanent deletion.
}
\examples{
delete_watchlist(watchlist_id = "ea3cc05b-2844-4985-b0c9-8f412718ac9e", live = FALSE)
}
\seealso{
Other Watchlist: 
\code{\link{add_to_watchlist}()},
\code{\link{delete_from_watchlist}()},
\code{\link{watchlist_update}()},
\code{\link{watchlist}()}
}
\concept{Watchlist}
