% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fetch_vars}
\alias{fetch_vars}
\title{fetch variables}
\usage{
fetch_vars(
  .vn,
  ...,
  evar = get0("evar", mode = "environment", envir = rlang::caller_env(), ifnotfound =
    new.env()),
  cenv = rlang::caller_env()
)
}
\arguments{
\item{.vn}{named vector of variable names with their corresponding classes - to be tested with \link[base]{inherits}}

\item{evar}{Environment with variables passed to top level of function}

\item{cenv}{The caller environment, stored automatically}

\item{e}{The list made from ellipsis if parameters are input as ellipsis argument}

\item{penv}{The parent of the caller environment, also stored automatically}

\item{sf}{The system frames for searching if not found in previous two environments.}
}
\description{
Fetches variables from environment of containing function for use in internal function
}
\keyword{internal}
