% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstRun.R
\name{firstrun}
\alias{firstrun}
\title{First Run Credential Setup}
\usage{
firstrun(paper, live)
}
\arguments{
\item{paper}{\code{(named character)} of length two: \code{key} and \code{secret} for the Alpaca paper API.}

\item{live}{\code{(named chacter)} of length two: \code{key} and \code{secret} for the Alpaca live API.}
}
\value{
Sets environment variables for this session and edits the .Renviron file in R's home directory \code{path.expand("~")} and adds the specified key/secret combinations for future sessions.
}
\description{
This function helps set-up your .Renviron file to work with \code{AlpacaforR}. It's for first time users and only needs to be run once.
}
\examples{
\dontrun{
firstrun(paper = c(key = "mypaperkey", secret = "mypapersecret"), live = c(key = "mylivekey", secret = "mylivesecret"))
}
}
