% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_account}
\alias{get_account}
\title{Get Account function}
\usage{
get_account(live = FALSE)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was specified.}
}
\value{
"id" Account ID as a string.

"status" Account Status as a string.

"currency" USD as a string.

"buying_power" Tradable buying power as a string.

"cash" Cash balance as a string.

"cash_withdrawable"  Withdrawable cash amount as a string.

"portfolio_value"  Total value of cash + holding positions as a string.

"trading_blocked"  If true, the account is not allowed to place orders as a boolean.

"transfers_blocked"  If true, the account is not allowed to request money transfers as a boolean.

"account_blocked"  If true, the account activity by user is prohibited as a boolean.

"created_at"  Timestamap this account was created at as a string.
}
\description{
The accounts API serves important information related to an account, including account status, funds available for trade, funds available for withdrawal, and various flags relevant to an account’s ability to trade.
}
\examples{
get_account(live = FALSE)
Which is similar to:
get_account()
For access to live accounts, you must submit as live = TRUE
get_account(live = TRUE)
}
