% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_assets}
\alias{get_assets}
\title{Get Assets function}
\usage{
get_assets(ticker = NULL, version = "v2")
}
\arguments{
\item{ticker}{Check for a specific stock and if its active on Alpaca by inputting ticker as a string.}
}
\value{
"id Asset" ID as a string.

"asset_class" us_equity as a string.

"exchange" AMEX, ARCA, BATS, NYSE, NASDAQ or NYSEARCA as a string.

"symbol" Stock symbol as a string.

"status" active or inactive as a string.

"tradeable" Asset is tradable on Alpaca or not as a boolean.
}
\description{
The assets API serves as the master list of assets available for trade and data consumption from Alpaca. Assets are sorted by asset class, exchange and symbol. Some assets are only available for data consumption via Polygon, and are not tradable with Alpaca. These assets will be marked with the flag tradable=false.
}
\examples{
get_assets()
# Get a specific asset:
get_assets(ticker = "AAPL",version = "v2")
}
