% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_bars}
\alias{get_bars}
\title{Get Bars function}
\usage{
get_bars(ticker, from = Sys.Date() - 7, to = NULL, timeframe = "1D",
  limit = 100)
}
\arguments{
\item{ticker}{The stock or stocks (in vector format) that you want.}

\item{from}{The starting date for the pricing data. Default is the last 5 trading days.}

\item{to}{The ending date for the pricing data. Default is NULL so it returns most recent data point.}

\item{timeframe}{One of "minute", "1Min", "5Min", "15Min", "day" or "1D". minute is an alias of 1Min. Similarly, day is of 1D. Defaults to "1D" as a string.}

\item{limit}{The amount of bars to return per ticker. This can range from 1 to 1000. Defaults to 100.}
}
\value{
"t" the beginning time of this bar as a Unix epoch in seconds as a integer.

"o" open price as a numberic object.

"h" high price as a numberic object.

"l" low price as a numberic object.

"c" close price as a numberic object.

"v" volume as a numberic object.
}
\description{
The bars API provides time-aggregated price and volume data for a single stock or multiple.
}
\examples{
get_bars(ticker = c("INTC","MSFT"))
get_bars(ticker = c("INTC","MSFT"), from = "2019-03-20", to = "2019-04-01", timeframe = "15Min", limit = 1000)
}
