% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_bars}
\alias{get_bars}
\title{Get Bars function}
\usage{
get_bars(
  ticker,
  from = Sys.Date() - 6,
  to = Sys.Date(),
  timeframe = "1D",
  limit = NULL
)
}
\arguments{
\item{ticker}{The stock or stocks (in vector format) that you want.}

\item{from}{The starting date for the pricing data. Default is the last 5 trading days.}

\item{to}{The ending date for the pricing data. Default is todays date.}

\item{timeframe}{One of "minute", "1Min", "5Min", "15Min", "day" or "1D". minute is an alias of 1Min. Similarly, day is of 1D. Defaults to "1D" as a string.}

\item{limit}{The amount of bars to return per ticker. This can range from 1 to 1000. Defaults according to timeframe chosen. If timeframe "1D or day" then the limit is set to the # of days. If "15Min" the default is 250, if "5Min" the default is 500, and if "1Min or minute" then the default is the max, 1000.}
}
\value{
\code{list} object for each ticker symbol containing a \code{data.frame} with the following columns:
\itemize{
 \item{\code{time}}{  the time of the bar as \code{POSIXct} in yyyy-mm-dd for timeframe = day, and yyyy-mm-dd hh:mm:ss for timeframes < day}
 \item{\code{open}}{  open price as a numeric object.}
 \item{\code{high}}{  high price as a numeric object.}
 \item{\code{low}}{  low price as a numeric object.}
 \item{\code{close}}{  close price as a numeric object.}
 \item{\code{volume}}{  volume (in millions) as a numeric object.}
}
}
\description{
The bars API provides time-aggregated price and volume data for a single stock or multiple.
}
\examples{
Getting one or more tickers: 
get_bars(ticker = c("INTC","MSFT"))
Getting price data with specific date ranges and timeframes, by also limiting the amount of bars returned for each ticker.
get_bars(ticker = c("INTC","MSFT"), from = "2019-03-20", to = "2019-04-01", timeframe = "15Min", limit = 175)
}
