% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_calendar}
\alias{get_calendar}
\title{Get Calendar function}
\usage{
get_calendar(from = NULL, to = NULL, version = "v2")
}
\arguments{
\item{from}{Starting date of request. Default to NULL so it starts from 1970.}

\item{to}{Ending date of request. Default to NULL so it ends in 2029.}
}
\value{
"date" Date string. in date format as a string.

"open" The time the market opens at on this date in hour:min format as a string.

"close" The time the market closes at on this date in hour:min format as a string.
}
\description{
The calendar API serves the full list of market days from 1970 to 2029. It can also be queried by specifying a start and/or end time to narrow down the results. In addition to the dates, the response also contains the specific open and close times for the market days, taking into account early closures.
}
\examples{
# Get all dates:
get_calendar(version = "v2")
# Get specific date range:
get_calendar(from = "2019-01-01", to = "2019-04-01", version = "v2")
}
