% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_clock}
\alias{get_clock}
\title{Get Clock function}
\usage{
get_clock(url, headers)
}
\arguments{
\item{url}{URL to your user live or paper API.}

\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}
}
\value{
"timestamp" Current timestamp as a string.

"is_open" Whether or not the market is open as a boolean.

"next_open" Next market open timestamp as a string.

"next_close" Next market close timestamp as a string.
}
\description{
The clock API serves the current market timestamp, whether or not the market is currently open, as well as the times of the next market open and close.
}
\examples{
get_clock(url = url, headers = headers)
}
