% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_headers}
\alias{get_headers}
\title{Get Headers for Server Request function}
\usage{
get_headers()
}
\value{
The correct headers that will be sent in the API request.
}
\description{
Get Headers for Server Request function
}
