% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R, R/alpaca4R.R
\name{get_meta}
\alias{get_meta}
\title{Get Polygon Meta Data (Deprecated)}
\usage{
get_meta(ticker = NULL, endpoint = NULL, perpage = NULL, version = "v1")

get_meta(ticker = NULL, endpoint = NULL, perpage = NULL, version = "v1")
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}

\item{endpoint}{Select either company for a company profile, analysts for all kinds of analyst estimates, dividends to view historic and upcoming dividends, earnings for historic and current earning details , or news for news updates from CNBC, Seeking Alpha, etc.}

\item{perpage}{This is only used if "news" was your selected endpoint. How many articles do you want to see perpage?}

\item{version}{The current version for API. Defaults to v1 if no v2 available.}

\item{v}{\code{(integer)} The Polygon API version number.}
}
\value{
A list object containing all information the API responds with.

A list object containing all information the API responds with.
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides more color on your stock through its available meta data endpoints from Polygon. These endpoints are company, analysts, dividends, earnings, and news.

This function provides more color on your stock through its available meta data endpoints from Polygon. These endpoints are company, analysts, dividends, earnings, and news.
}
\examples{
# Getting default meta for AMZN: 
get_meta(ticker = "AMZN")
# Getting news information on AMZN: 
get_meta(ticker = "AMZN", endpoint = "news", perpage = 100)
# Getting default meta for AMZN: 
get_meta(ticker = "AMZN")
# Getting news information on AMZN: 
get_meta(ticker = "AMZN", endpoint = "news", perpage = 100)
}
\seealso{
Other Polygon: 
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()},
\code{\link{polygon}()}
}
\concept{Polygon}
