% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_orders}
\alias{get_orders}
\title{Get Orders function}
\usage{
get_orders(url, headers, status, from = NULL)
}
\arguments{
\item{url}{URL to your user live or paper API.}

\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}

\item{status}{Order status to be queried "open, closed or all". Defaults to open as a string.}

\item{from}{The response will include only orders submitted after this date exclusive as a timestamp object.}
}
\value{
"id" order id as a string.

"client_order_id" client unique order id as a string.

"created_at" nullable When the order was created as a timestamp object.

"updated_at" nullable When the order was created as a timestamp object.

"submitted_at" nullable When the order was created as a timestamp object.

"filled_at" nullable When the order was created as a timestamp object.

"expired_at" nullable When the order was created as a timestamp object.

"canceled_at" nullable When the order was created as a timestamp object.

"asset_id" asset ID as a string.

"symbol" Asset symbol as a string.

"exchange" Asset exchange as a string.

"asset_class" Asset class as a string.

"qty" Ordered quantity as a string.

"filled_qty" Filled quantity as a string.

"type" Valid values: market, limit, stop, stop_limit as a string.

"side" Valid values: buy, sell as a string.

"time_in_force" time in force selected asa string.

"limit_price" Limit price as a string.

"stop_price" Stop price as a string.

"status" Status of the order as a string.
}
\description{
The orders API allows a user to monitor, place and cancel their orders with Alpaca.
}
\examples{
get_orders(url = url, headers = headers, status = "open")
}
