% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{get_poly_historic_info}
\alias{get_poly_historic_info}
\title{Get Polygon Historic Trades Information(Deprecated)}
\usage{
get_poly_historic_info(
  ticker = NULL,
  type = NULL,
  date = NULL,
  .ts = NULL,
  v = 2,
  limit = 10
)
}
\arguments{
\item{ticker}{\code{(character)} Specify which symbol you want.}

\item{type}{\code{(character)} Get information on either "t/trades" or "q/quotes".}

\item{date}{\code{(character/Date/POSIXct)} Specify the date for which you are requesting in YYYY-MM-DD format.}

\item{.ts}{\code{(Datetime)} Timestamp offset, used for pagination. This is the offset at which to start the results. Using the timestamp of the last result as the offset will give you the next page of results.}

\item{limit}{\code{(integer)} Limit the size of response, Max 50000. Default 10.}
}
\value{
\code{(data.frame)} Object containing all information specified in the Properties section of the Endpoint links above with the following attributes:

\code{(list)} Accessed via \code{attr(result, "query")} with query metadata as follows:
\itemize{
\item{\code{success} \code{(logical)} If this query was executed successfully}
\item{\code{ticker} \code{(character)} Ticker symbol that was evaluated from the request}
\item{\code{results_count} \code{(integer)} Total number of results in this response}
\item{\code{db_latency} \code{(integer)} Milliseconds of latency for the query results from DB}
}

\code{(list)} Accessed via \code{attr(result, "map")}. The \code{map} of the results (identifiers and classes as returned by the API). `
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides historic trade data form Polygon on the ticker specified. It returns a list with values such as date, bid size / ask size, the exchanges for bid / ask, latency, each trade/quote listed for that date, time, etc. See the \url{Polygon.io/docs/}: \href{https://polygon.io/docs/#get_v2_ticks_stocks_trades__ticker___date__anchor}{Historic Trades Endpoint} & \href{https://polygon.io/docs/#get_v2_ticks_stocks_nbbo__ticker___date__anchor}{Historic Quotes Endpoint}
}
\examples{
# Getting historic trade data on AMZN: 
get_poly_historic_info(ticker = "AMZN", type = "trades", date = "2019-04-05")
# Getting historic pricing data on AMZN: 
get_poly_historic_info(ticker = "AMZN", type = "quotes", date = "2019-04-05")
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()},
\code{\link{polygon}()}
}
\concept{Polygon}
