% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_historic_info}
\alias{get_poly_historic_info}
\title{Get Polygon Historic Trades Information}
\usage{
get_poly_historic_info(ticker = NULL, type = NULL, date = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}

\item{type}{Get informaiton on either "trades" or "quotes".}

\item{date}{Specify the date for which you are requesting.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides historic trade data form Polygon on the ticker specified. It returns a list with values such as date, bid size / ask size, the exchanges for bid / ask, latency, each trade/quote listed for that date, time, etc.
}
\examples{
# Getting historic trade data on AMZN: 
get_poly_historic_info(ticker = "AMZN", type = "trades", date = "2019-04-05")
# Getting historic pricing data on AMZN: 
get_poly_historic_info(ticker = "AMZN", type = "quotes", date = "2019-04-05")
}
