% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_last_price}
\alias{get_poly_last_price}
\title{Get Polygon Last Price}
\usage{
get_poly_last_price(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides the last listed price from Polygon. A list is returned with values such as the last price, last size, last exchange, and last timestamp.
}
\examples{
# Getting the last listed price for AMZN: 
get_poly_last_price("AMZN")
}
