% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_last_trade}
\alias{get_poly_last_trade}
\title{Get Polygon Last Trade}
\usage{
get_poly_last_trade(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides the last listed trade from Polygon for the ticker specified. A list is returned with values such as the last bid / ask price, bid / ask size, bid / ask exchange, and last trade timestamp.
}
\examples{
# Getting the last listed trade for AMZN: 
get_poly_last_trade("AMZN")
}
