% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_ohlc}
\alias{get_poly_ohlc}
\title{Get Polygon Daily OHLCV & After Hours}
\usage{
get_poly_ohlc(ticker = NULL, date = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}

\item{date}{Specify the date for which you are requesting.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides the last OHLCV data, including after hours data from Polygon for the ticker specified.
}
\examples{
Getting the last listed trade for AMZN: 
get_poly_ohlc("AMZN", date = "2019-03-20")
}
