% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R, R/alpaca4R.R
\name{get_poly_stock_splits}
\alias{get_poly_stock_splits}
\title{Get Polygon Stock Split Information (Deprecated)}
\usage{
get_poly_stock_splits(ticker = NULL)

get_poly_stock_splits(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.

A list object containing all information the API responds with.
}
\description{
Deprecated. See \code{\link[AlpacaforR]{polygon}}.
This function provides stock split data for the specified ticker from Polygon.

This function provides stock split data for the specified ticker from Polygon.
}
\examples{
# Getting default pricing data on AMZN (daily): 
get_poly_stock_splits(ticker = "AMZN")
# Getting default pricing data on AMZN (daily): 
get_poly_stock_splits(ticker = "AMZN")
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{polygon}()}
}
\concept{Polygon}
