% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_poly_stock_splits}
\alias{get_poly_stock_splits}
\title{Get Polygon Stock Split Information}
\usage{
get_poly_stock_splits(ticker = NULL)
}
\arguments{
\item{ticker}{Specify which symbol you want to call by inserting ticker as a string.}
}
\value{
A list object containing all information the API responds with.
}
\description{
This function provides stock split data for the specified ticker from Polygon.
}
\examples{
Getting default pricing data on AMZN (daily): 
get_poly_stock_splits(ticker = "AMZN")
}
