% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_positions}
\alias{get_positions}
\title{Get Positions function}
\usage{
get_positions(url, headers, ticker = NULL)
}
\arguments{
\item{url}{URL to your user live or paper API.}

\item{headers}{Header object containing your KeyID & SecretID to your user API. Use add_headers('APCA-API-KEY-ID' = value, 'APCA-API-SECRET-KEY' = value) from httr package.}
}
\value{
"asset_id"  Asset ID as a string.

"symbol"  Symbol of the asset as a string.

"exchange"  Exchange name of the asset as a string.

"asset_class"  Asset class name as a string.

"avg_entry_price"  Average entry price of the position as a string.

"qty" The number of shares as a string.

"side" long/short as a string.

"market_value"  Total dollar amount of the position as a string.

"cost_basis"  Total cost basis in dollar as a string.

"unrealized_pl"  Unrealized profit/loss in dollar as a string.

"unrealized_plpc"  Unrealized profit/loss percent (by a factor of 1) as a string.

"unrealized_intraday_pl"  Unrealized profit/loss in dollar for the day as a string.

"unrealized_intraday_plpc"  Unrealized profit/loss percent (by a factor of 1) as a string.

"current_price"  Current asset price per share as a string.

"lastday_price"  Last day’s asset price per share as a string.

"change_today"  Percent change from last day price (by a factor of 1) as a string.
}
\description{
The positions API provides information about an account’s current open positions. The response will include information such as cost basis, shares traded, and market value, which will be updated live as price information is updated.
}
\examples{
get_positions(url = url, headers = headers) 
}
