% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{get_url}
\alias{get_url}
\title{Get URL for Server Request function}
\usage{
get_url(paper)
}
\arguments{
\item{paper}{A parameter used in functions that require knowledge of the account type.}
}
\value{
The correct URL according to account type (live or paper) that will be sent in the API request.
}
\description{
Get the correct URL for the Server Request that is sent to interact with the API. If the user is on a paper account, then the paper account URL will be returned.
}
