% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{\%|z|\%}
\alias{\%|z|\%}
\title{Default value for zero length variable}
\usage{
x %|z|% y
}
\arguments{
\item{x, y}{If \code{x} is length zero, will return \code{y}; otherwise returns \code{x}}
}
\description{
This infix replaces zero length variables with a default value
}
\examples{
numeric(0) \%|z|\% NA_real_
}
