% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_id}
\alias{is_id}
\title{Check if value provided to a ticker_id argument is an ID or a ticker symbol}
\usage{
is_id(ticker_id)
}
\arguments{
\item{ticker_id}{}
}
\value{
\code{logical} indicating whether the object is a
}
\description{
for use in functions that accept \code{ticker_id}
}
\keyword{internal}
