% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{match_letters}
\alias{match_letters}
\title{Match the first x letters to supplied arguments}
\usage{
match_letters(arg, ..., x = 1, several.ok = FALSE, capitalize = FALSE)
}
\description{
Case insensitive matching of argument to possibilities provided in ellipsis.
}
