% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{order_symbol_id}
\alias{order_symbol_id}
\title{retrieve the order id if an order object is supplied}
\usage{
order_symbol_id(symbol_id, side, action, qty, client_order_id)
}
\arguments{
\item{symbol_id}{\code{(character)}  The stock symbol (\emph{Required} when \code{action = "submit"}) or Order object (single row tibble) (\emph{Required} when \code{action = "cancel"/"replace"}).
To expedite the setting of stops and limits for open positions, an Order ID from a \code{'buy'} order can be provided when \code{action = "submit"} to place a \code{'sell'} order with the following parameters such that they do not need to be set manually:
\itemize{
\item{\code{side = 'sell'}}
\item{If \code{qty} is not provided, it will be populated from the buy order}
\item{\code{symbol_id} will be set to the symbol from the buy order.}
\item{If \code{client_order_id = TRUE}, the \code{client_order_id} will be set to the buy Order ID provided, effectively linking the orders for your records.}
\item{All other parameters can be specified as usual.}
}}

\item{side}{\code{(character)} The side of the trade. I.E \code{"buy"/"b"} or \code{"sell"/"s"}. (\emph{required} when \code{action = "submit"}). Assumed to be \code{"buy"} if \code{order_class = "bracket"}.}

\item{action}{\code{(character)} The action to take:
\itemize{
\item{\code{"submit"/"s"}}{ \href{https://alpaca.markets/docs/api-documentation/api-v2/orders/#request-a-new-order}{Request a new order} \strong{Default}}
\item{\code{"replace"/"r"}}{ \href{https://alpaca.markets/docs/api-documentation/api-v2/orders/#replace-an-order}{Replace an order}}
\item{\code{"cancel"/"c"}}{ \href{https://alpaca.markets/docs/api-documentation/api-v2/orders/#cancel-an-order}{Cancel an order}}
\item{\code{"cancel_all"}}{ \href{https://alpaca.markets/docs/api-documentation/api-v2/orders/#cancel-all-orders}{Cancel all orders}}
}}

\item{qty}{\code{(integer)} The amount of shares to trade (\emph{required} when \code{action = "submit"}, \emph{optional} when \code{action = 'replace'}).}

\item{client_order_id}{\code{(character/logical)}  <= 48 Characters.  A unique identifier for the order. Automatically generated if not sent.
\itemize{
\item{\code{`action = 'replace'/'submit'`}}{ \emph{Optional} character vector}
\item{\code{`action = 'submit'`}}{ If an Order object is provided to \code{symbol_id}, \code{TRUE} will set the \code{client_order_id} for the sell order to Order ID in \code{symbol_id}. Used to link buy & sell orders for your records.}
}}
}
\description{
Retrieves the order_id from an order object if provided as \code{symbol_id}, or provides informative error if input order failed.
}
\keyword{internal}
