% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orders.R
\name{submit_orders}
\alias{submit_orders}
\title{submit_orders}
\usage{
submit_orders(
  ticker,
  qty,
  side,
  type = NULL,
  time_in_force = "day",
  limit_price = NULL,
  stop_price = NULL,
  extended_hours = FALSE,
  live = FALSE,
  v = 2
)
}
\description{
\code{submit_orders} is deprecated. Use \code{\link[AlpacaforR]{order_submit}} instead.
}
\examples{
submit_orders()
}
\seealso{
Other Orders: 
\code{\link{order_cancel}()},
\code{\link{order_replace}()},
\code{\link{order_submit}()},
\code{\link{orders_get}()}
}
\concept{Orders}
