% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{poly_transform}
\alias{poly_transform}
\title{transform polygon.io reference endpoints}
\usage{
poly_transform(resp, ep)
}
\arguments{
\item{resp}{The response object from httr}

\item{ep}{The endpoint}
}
\value{
\code{list/data.frame/tibble} Either a list or tibble depending on the endpoint
}
\description{
transform polygon.io reference endpoints
}
\keyword{R}
\keyword{coerce}
\keyword{compliant}
\keyword{data}
\keyword{endpoints}
\keyword{from}
\keyword{internal}
\keyword{objects}
\keyword{polygon.io}
\keyword{returned}
\keyword{the}
\keyword{to}
\keyword{various}
