% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{poly_transform}
\alias{poly_transform}
\title{transform polygon.io reference endpoints}
\usage{
poly_transform(resp, e_p)
}
\arguments{
\item{resp}{\code{\link[httr]{response}} from \code{{httr}}}

\item{e_p}{\code{(list)} The endpoint description}
}
\value{
\code{list/data.frame/tibble} Either a list or tibble depending on the endpoint with a query attribute.
}
\description{
coerce data objects returned from the various polygon.io endpoints to R compliant objects
}
\keyword{internal}
