% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon.R
\name{polygon}
\alias{polygon}
\title{Access all Polygon.io Endpoints}
\usage{
polygon(ep = NULL, ..., params = NULL)
}
\arguments{
\item{ep}{\code{(character)} Non-case-sensitive reference to the endpoint to be called as an abbreviation or by name. See Details for full endpoint details. Use the following keywords to quick-view abbreviations and full names for endpoints while programming (add a '+' to any keyword to return a list with full details: abbv, name, url, description, parameters etc):
\itemize{
\item{\code{'all'}}{ View all endpoints that Alpaca users have access to.}
\item{\code{'ref'/'reference'}}{ View all reference endpoints.}
\item{\code{'sto'/'stocks'}}{ View all stock endpoints}
}
Alternatively, if you know how to call the endpoint but would like a list of it's details, add '+' to the end of it (ie 'tt+' or 'ticker types+'.}

\item{...}{Named arguments specific to the endpoint parameters. These can also be specified in a list via the \code{params} argument.}

\item{params}{\code{(list)} A named list of parameters specific to the endpoint.}
}
\value{
Response \code{(tibble/list/data.frame)} depending on the endpoint. The core data of the response will be returned as the object. If query data is returned in addition to the core object, it is stored as a \code{"query"} attribute and accessed via \code{attr(object, "query")}. If a map object is returned with the response, it will be stored as a \code{"map"} attribute and accessed via \code{attr(object, "map")}.
}
\description{
This function calls all Endpoints from the \href{https://polygon.io/docs}{Polygon API docs}.
}
\details{
Optional query parameters are mentioned if present. See the \href{https://polygon.io/docs}{Polygon API docs} for details on parameters and response data. Required path parameters are in \strong{bold}. The first option for each required parameter is used as a default if none are specified. Endpoints are:
\itemize{
\item{\code{t}}{ \href{https://polygon.io/docs/#get_v2_reference_tickers_anchor}{Tickers}: Query all ticker symbols which are supported by Polygon.io. This API includes Indices, Crypto, FX, and Stocks/Equities.
Arguments:
\itemize{
\item{\code{"sort"}}{ \code{(character)} ticker, -ticker, type}
\item{\code{"type"}}{ \code{(character)} \emph{Optional} NULL, etp, cs}
\item{\code{"market"}}{ \code{(character)} \emph{Optional} NULL, stocks, indices}
\item{\code{"locale"}}{ \code{(character)} \emph{Optional} NULL, us, g}
\item{\code{"search"}}{ \code{(character)} \emph{Optional} NULL, microsoft}
\item{\code{"perpage"}}{ \code{(numeric)} 50}
\item{\code{"page"}}{ \code{(numeric)} 1}
\item{\code{"active"}}{ \code{(logical)} \emph{Optional} NULL, TRUE, FALSE}
}
}
\item{\code{tt}}{ \href{https://polygon.io/docs/#get_v2_reference_types_anchor}{Ticker Types}: Get the mapping of ticker types to descriptions / long names
}
\item{\code{td}}{ \href{https://polygon.io/docs/#get_v1_meta_symbols__symbol__company_anchor}{Ticker Details}: Get the details of the symbol company/entity. These are important details which offer an overview of the entity. Things like name, sector, description, logo and similar companies.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
}
}
\item{\code{tn}}{ \href{https://polygon.io/docs/#get_v1_meta_symbols__symbol__news_anchor}{Ticker News}: Get news articles for this ticker.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
\item{\code{"perpage"}}{ \code{(numeric)} 50}
\item{\code{"page"}}{ \code{(numeric)} 1}
}
}
\item{\code{m}}{ \href{https://polygon.io/docs/#get_v2_reference_markets_anchor}{Markets}: Get the list of currently supported markets
}
\item{\code{l}}{ \href{https://polygon.io/docs/#get_v2_reference_locales_anchor}{Locales}: Get the list of currently supported locales
}
\item{\code{ss}}{ \href{https://polygon.io/docs/#get_v2_reference_splits__symbol__anchor}{Stock Splits}: Get the historical splits for this symbol.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
}
}
\item{\code{sd}}{ \href{https://polygon.io/docs/#get_v2_reference_dividends__symbol__anchor}{Stock Dividends}: Get the historical dividends for this ticker.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
}
}
\item{\code{sf}}{ \href{https://polygon.io/docs/#get_v2_reference_financials__symbol__anchor}{Stock Financials}: Current status of each market
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
\item{\code{"limit"}}{ \code{(numeric)} 5}
\item{\code{"type"}}{ \code{(character)} \emph{Optional} NULL, Y, YA, Q, QA, T, TA}
\item{\code{"sort"}}{ \code{(character)} \emph{Optional} NULL, reportPeriod, -reportPeriod, calendarDate, -calendarDate}
}
}
\item{\code{ms}}{ \href{https://polygon.io/docs/#get_v1_marketstatus_now_anchor}{Market Status}: Current status of each market
}
\item{\code{mh}}{ \href{https://polygon.io/docs/#get_v1_marketstatus_upcoming_anchor}{Market Holidays}: Get upcoming market holidays and their open/close times
}
\item{\code{e}}{ \href{https://polygon.io/docs/#get_v1_meta_exchanges_anchor}{Exchanges}: List of stock exchanges which are supported by Polygon.io
}
\item{\code{ht}}{ \href{https://polygon.io/docs/#get_v2_ticks_stocks_trades__ticker___date__anchor}{Historic Trades}: Get historic NBBO quotes for a ticker.
Arguments:
\itemize{
\item{\strong{\code{ticker}}}{ \code{(character)} AAPL}
\item{\strong{\code{date}}}{ \code{(Date)} 2018-02-02}
\item{\code{"timestamp"}}{ \code{(numeric)} \emph{Optional} NULL, 1}
\item{\code{"timestampLimit"}}{ \code{(numeric)} \emph{Optional} NULL, 1}
\item{\code{"reverse"}}{ \code{(logical)} \emph{Optional} NULL, TRUE, FALSE}
\item{\code{"limit"}}{ \code{(numeric)} 10, 50000}
}
}
\item{\code{hq}}{ \href{https://polygon.io/docs/#get_v2_ticks_stocks_nbbo__ticker___date__anchor}{Historic Quotes (NBBO)}: Get historic NBBO quotes for a ticker.
Arguments:
\itemize{
\item{\strong{\code{ticker}}}{ \code{(character)} AAPL}
\item{\strong{\code{date}}}{ \code{(Date)} 2018-02-02}
\item{\code{"timestamp"}}{ \code{(numeric)} \emph{Optional} NULL, 1}
\item{\code{"timestampLimit"}}{ \code{(numeric)} \emph{Optional} NULL, 1}
\item{\code{"reverse"}}{ \code{(logical)} \emph{Optional} NULL, TRUE, FALSE}
\item{\code{"limit"}}{ \code{(numeric)} 10, 50000}
}
}
\item{\code{lt}}{ \href{https://polygon.io/docs/#get_v1_last_stocks__symbol__anchor}{Last Trade for a Symbol}: Get the last trade for a given stock.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
}
}
\item{\code{lq}}{ \href{https://polygon.io/docs/#get_v1_last_quote_stocks__symbol__anchor}{Last Quote for a Symbol}: Get the last quote tick for a given stock.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
}
}
\item{\code{do}}{ \href{https://polygon.io/docs/#get_v1_open-close__symbol___date__anchor}{Daily Open/Close}: Get the open, close and afterhours prices of a symbol on a certain date.
Arguments:
\itemize{
\item{\strong{\code{symbol}}}{ \code{(character)} AAPL}
\item{\strong{\code{date}}}{ \code{(Date)} 2020-05-04}
}
}
\item{\code{cm}}{ \href{https://polygon.io/docs/#get_v1_meta_conditions__ticktype__anchor}{Condition Mappings}: The mappings for conditions on trades and quotes.
Arguments:
\itemize{
\item{\strong{\code{ticktype}}}{ \code{(character)} trades, quotes}
}
}
\item{\code{sa}}{ \href{https://polygon.io/docs/#get_v2_snapshot_locale_us_markets_stocks_tickers_anchor}{Snapshot - All tickers}: Snapshot allows you to see all tickers current minute aggregate, daily aggregate and last trade. As well as previous days aggregate and calculated change for today.
}
\item{\code{st}}{ \href{https://polygon.io/docs/#get_v2_snapshot_locale_us_markets_stocks_tickers__ticker__anchor}{Snapshot - Single Ticker}: See the current snapshot of a single ticker
Arguments:
\itemize{
\item{\strong{\code{ticker}}}{ \code{(character)} AAPL}
}
}
\item{\code{sg}}{ \href{https://polygon.io/docs/#get_v2_snapshot_locale_us_markets_stocks__direction__anchor}{Snapshot - Gainers/Losers}: See the current snapshot of the top 20 gainers or losers of the day at the moment.
Arguments:
\itemize{
\item{\strong{\code{direction}}}{ \code{(character)} gainers, losers}
}
}
\item{\code{pc}}{ \href{https://polygon.io/docs/#get_v2_aggs_ticker__ticker__prev_anchor}{Previous Close}: Get the previous day close for the specified ticker
Arguments:
\itemize{
\item{\strong{\code{ticker}}}{ \code{(character)} AAPL}
\item{\code{"unadjusted"}}{ \code{(logical)} \emph{Optional} NULL, TRUE, FALSE}
}
}
\item{\code{a}}{Aggregates - see ?market-data}
\item{\code{gd}}{ \href{https://polygon.io/docs/#get_v2_aggs_grouped_locale__locale__market__market___date__anchor}{Grouped Daily (Bars)}: Get the daily OHLC for entire markets.
Arguments:
\itemize{
\item{\strong{\code{locale}}}{ \code{(character)} US, See Locales Endpoint}
\item{\strong{\code{market}}}{ \code{(character)} STOCKS, See Markets Endpoint}
\item{\strong{\code{date}}}{ \code{(Date)} 2019-02-01}
\item{\code{"unadjusted"}}{ \code{(logical)} \emph{Optional} NULL, TRUE, FALSE}
}
}
}

This function is not vectorized. Only a single endpoint may be called at a time. Thus any endpoints with path parameters (parameters denoted in bold) require that a single combination of path parameters be passed for a given call.
}
\seealso{
Other Polygon: 
\code{\link{get_meta}()},
\code{\link{get_poly_agg_quote}()},
\code{\link{get_poly_historic_info}()},
\code{\link{get_poly_last_price}()},
\code{\link{get_poly_last_trade}()},
\code{\link{get_poly_ohlc}()},
\code{\link{get_poly_prev_dayclose}()},
\code{\link{get_poly_stock_splits}()}
}
\concept{Polygon}
