% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Positions.R
\name{positions}
\alias{positions}
\alias{get_positions}
\alias{close_all_positions}
\title{Get Positions}
\usage{
positions(tickers = NULL, action = "get", live = FALSE, v = 2)

get_positions(ticker = NULL, live = FALSE, version = "v2")

close_all_positions(live = FALSE)
}
\arguments{
\item{tickers}{\code{(character)} Specify which symbol(s) you want to call by inserting ticker symbol(s) as a character vector. Use \code{"close_all"} to close all positions.}

\item{action}{\code{(character)} \code{"get"/"g"} to get all positions or those specified by \code{tickers}. \verb{"close","c"} to close positions specified by \code{tickers}. Use \code{tickers = "close_all"} to close all positions - \code{action} is automatically set to \code{"close"} when \code{tickers = "close_all"}.}

\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}

\item{v}{\code{(integer)} The API version number. Defaults to 2. There is only the V2 API available for all endpoints other than Bars as of 2020-04-18. This argument is present to accommodate future API versions.}
}
\value{
Position \code{(tibble)} \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/#position-entity}{Position object} or array of Position objects of length 0 if no positions, otherwise with length of the number of positions, each with the following attributes:
\itemize{
\item{\code{asset_id}}{\code{(character)} Asset ID.}
\item{\code{symbol}}{\code{(character)} Symbol of the asset.}
\item{\code{exchange}}{\code{(character)} Exchange name of the asset.}
\item{\code{asset_class}}{\code{(character)} Asset class name.}
\item{\code{qty}}{\code{(integer)} The number of shares.}
\item{\code{avg_entry_price}}{\code{(numeric)} Average entry price of the position.}
\item{\code{side}}{\code{(character)} long/short exposure.}
\item{\code{market_value}}{\code{(numeric)} Total dollar amount of the position.}
\item{\code{cost_basis}}{\code{(numeric)} Total cost basis in dollar.}
\item{\code{unrealized_pl}}{\code{(numeric)} Unrealized profit/loss in dollar.}
\item{\code{unrealized_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{unrealized_intraday_pl}}{\code{(numeric)} Unrealized profit/loss in dollar for the day.}
\item{\code{unrealized_intraday_plpc}}{\code{(numeric)} Unrealized profit/loss percent (by a factor of 1).}
\item{\code{current_price}}{\code{(numeric)} Current asset price per share.}
\item{\code{lastday_price}}{\code{(numeric)} Last day's asset price per share.}
\item{\code{change_today}}{\code{(numeric)} Percent change from last day price (by a factor of 1).}
}
}
\description{
The positions API provides information about an account’s current open positions. The response will include information such as cost basis, shares traded, and market value, which will be updated live as price information is updated. Character values are returned as a string while integer values are returned as numeric. See \href{https://alpaca.markets/docs/api-documentation/api-v2/positions/}{Positions} for details.

\code{get_positions} is deprecated, use \code{\link[AlpacaforR]{positions}} instead.

Deprecated. Please use \link[AlpacaforR]{\code{positions}} with `tickers = 'close_all'". The close all positions API liquidates all currently open long and short positions. Character values are returned as a string while integer values are returned as numeric.
}
\examples{
positions(ticker = "AAPL", live = FALSE)
positions(ticker = "AAPL", live = TRUE)
positions() # all paper positions
positions(live = TRUE) # all live positions
get_positions()
close_all_positions(live = FALSE)
close_all_positions()
close_all_positions(live = TRUE)
}
\seealso{
Other Positions: 
\code{\link{close_position}()}

Other Positions: 
\code{\link{close_position}()}

Other Positions: 
\code{\link{close_position}()}
}
\concept{Positions}
