% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{response_text_clean}
\alias{response_text_clean}
\title{Clean Text from Server Response function}
\usage{
response_text_clean(dat)
}
\arguments{
\item{dat}{The response from our server GET request usually in a json format.}
}
\description{
Clean the response text (usually in unreadable json) and convert to a readable format using fromJSON.
}
