% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca4R.R
\name{submit_order}
\alias{submit_order}
\title{Submit Order function}
\usage{
submit_order(live = FALSE, ticker, qty, side, type,
  time_in_force = "day", limit_price = NULL, stop_price = NULL)
}
\arguments{
\item{live}{TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was specified.}

\item{ticker}{The stock's symbol as a string.}

\item{qty}{The amount of shares to trade as a string.}

\item{side}{The side of the trade. I.E buy or sell as a string.}

\item{type}{The type of trade order. I.E market,limit,stop,stoplimit,etc as a string.}

\item{time_in_force}{The type of time order. I.E day, gtc, etc as a string. Defaults to "day".}

\item{limit_price}{If order type was a limit, then enter the limit price here as a string.}

\item{stop_price}{If order tyope was a stop, then enter the stop price here as a string.}
}
\description{
Places a new order of the specified stock, quantity, buy / sell, type of order, time in force, and limit / stop prices if selected.
}
\examples{
For market order:
submit_order(ticker = "AAPL", qty = "100", side = "buy", type = "market")
Or you can submit a limit order:
submit_order(ticker = "AAPL", qty = "100", side = "buy", type = "limit", limit_price = "120")
}
