% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{time_index}
\alias{time_index}
\title{time_index}
\usage{
time_index(x, type = "character")
}
\arguments{
\item{type}{\code{(character)} If TRUE
\itemize{
\item{\code{"name"/"n"}}{ (Default) return the name of the index}
\item{\code{"value"/"v"}}{ return the vector of values in the time index}
\item{\code{"symbol"/"s"}}{ return the symbol for the time index}
}}

\item{.d}{\code{(data.frame/xts)} input data}
}
\value{
\code{(character/numeric)} Either the name of the index, or the index itself based on \code{name} argument
}
\description{
Get the time index for a data.frame or xts
}
