% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{try_date}
\alias{try_date}
\title{String to Date/Datetime conversion for market_data}
\usage{
try_date(.x)
}
\description{
Attempts to coerce string to a valid date object in stock exchange time zone, or checks to see if it is already. If it fails to coerce to date, returns NA.
}
\keyword{internal}
