% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-verbs.R
\name{ungroup.tsymble_ts}
\alias{ungroup.tsymble_ts}
\title{ungroup.tsymble_ts}
\usage{
\method{ungroup}{tsymble_ts}(x, ...)
}
\arguments{
\item{x}{A \code{\link[dplyr:tbl]{tbl()}}}

\item{...}{In \code{group_by()}, variables or computations to group by.
In \code{ungroup()}, variables to remove from the grouping.}
}
\value{
A grouped data frame with class \code{\link[dplyr]{grouped_df}},
unless the combination of \code{...} and \code{add} yields a empty set of
grouping columns, in which case a tibble will be returned.
}
\description{
An S3 method for ungrouping a \code{tsymble_ts} object
}
