% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watchlist.R
\name{watchlist}
\alias{watchlist}
\title{Access all Watchlist Endpoints}
\usage{
watchlist(
  watchlist_id = NULL,
  name = NULL,
  tickers = NULL,
  action = NULL,
  live = as.logical(Sys.getenv("APCA-LIVE", FALSE))
)
}
\arguments{
\item{watchlist_id}{\code{(character)} \strong{Name or ID} of watchlist to \code{GET}, \code{PATCH} or \code{DELETE}. If \code{NULL} \strong{(Default)} the array of all saved \code{Watchlist} will be returned. If \code{NULL} and \code{name} is specified, \code{action} is assumed to be \code{"create"} and a \code{Watchlist} named \code{name} will be created with any \code{tickers} specified. All other actions to be performed on the provided \code{Watchlist} name or ID must be explicitly specified via \code{action}. See other parameters for details. The list below describes the behavior of this argument based on the value supplied to \code{action}:
\itemize{
\item{\code{NULL}}{ If \code{watchlist_id = NULL} & \code{action = NULL}, and no other arguments specified. An array of all \code{Watchlist} will be returned. If \code{name} is specified, a \code{Watchlist} by that name will be created.}
\item{\code{"create"}}{ A \code{Watchlist} by this name will be created. If \code{watchlist_id} & \code{name} are specified the \code{Watchlist} specified with \code{watchlist_id} will be renamed to \code{name} preserving existing tickers.}
\item{\code{"add"}}{ If \code{tickers} are specified they will be added to the \code{Watchlist} specified.}
\item{\code{"replace"}}{ If \code{tickers} are specified they will replace existing tickers or if \code{tickers = NULL}, all tickers will be removed.}
\item{\code{"delete"}}{ If no tickers are specified, the \code{Watchlist} specified by \code{watchlist_id} will be deleted. If \code{tickers} are specified, the \code{Watchlist} specified by \code{watchlist_id} will have the specified tickers deleted from it.}
}}

\item{name}{\code{(character)} arbitrary name string, up to 64 characters. Default \code{NULL}. If only this is provided, \code{action} is assumed to be \code{"create"}. If only \code{name} & \code{watchlist_id} is provided the \code{Watchlist} specified by \code{watchlist_id} will be renamed to \code{name} preserving existing tickers. If this argument is specified and \code{action} is:
\itemize{
\item{\code{"create"/NULL}}{ A \code{Watchlist} by this name will be created, unless \code{watchlist_id} is specified, in which case the specified \code{Watchlist} will be renamed.}
\item{\code{"add"}}{ The \code{watchlist_id} specified will be renamed to this \code{name}. If \code{tickers} are specified they will be added or existing tickers will be preserved if \code{tickers = NULL}.}
\item{\code{"replace"}}{ The \code{watchlist_id} specified will be named to this \code{name}. If \code{tickers} are specified they will replace existing tickers or if \code{tickers = NULL}, all tickers will be removed.}
\item{\code{"delete"}}{ If no tickers are specified, this argument will be ignored and the \code{Watchlist} specified by \code{watchlist_id} will be deleted. If tickers are specified, the \code{Watchlist} specified by \code{watchlist_id} will be renamed to this \code{name} and the specified tickers will be deleted from it.}
}}

\item{tickers}{\code{(character)} Vector of symbols. Must be specified with a \code{watchlist_id} or \code{name}. Default \code{NULL}. If the \code{action} is:
\itemize{
\item{\code{"create"/ NULL}}{ If this argument is specified with \code{name} a \code{Watchlist} will be created with these symbols. If \code{NULL} the \code{Watchlist} will be empty.}
\item{\code{"add"}}{ If this argument is specified with \code{watchlist_id} the \code{tickers} will be added to the specified \code{Watchlist}. If this argument is specified with \code{watchlist_id} and \code{name} the \code{Watchlist} will be renamed and the \code{tickers} will be added.}
\item{\code{"replace"}}{ If this argument is specified with \code{watchlist_id} the \code{tickers} will replace those in the specified \code{Watchlist}. If this argument is specified with \code{watchlist_id} and \code{name} the \code{Watchlist} will be renamed and the \code{tickers} will replace the existing ones. If \code{NULL} all tickers will be removed and the \code{Watchlist} will be empty.}
\item{\code{"delete"}}{ If this argument is specified with \code{watchlist_id}, the \code{tickers} will be deleted from the \code{Watchlist}. If a \code{name} is also specified, the \code{Watchlist} will also be renamed. If \code{tickers} is \code{NULL} the \code{Watchlist} will be deleted.}
}}

\item{action}{\code{(character)} The action to take. See other parameters for detailed descriptions of how this argument determines the action when combinations of parameters are specified. This argument can be abbreviated with a single non-case-sensitive letter. One of:
\itemize{
\item{\code{NULL}}{ If \code{action = NULL} & \code{watchlist_id = NULL}, and no other arguments specified. An array of all \code{Watchlist} will be returned. If only \code{name} is specified, \code{action} is assumed to be \code{create}. If \code{tickers} are specified, the tickers will be added.}
\item{\code{"create"/"c"}}{\code{name} must be provided, and the a list will be created by this \code{name}. This action is assumed when only \code{name} is provided and \code{action = NULL}}
\item{\code{"add"/"a"}}{ If \code{"add"}, the specified \code{watchlist_id} will have \code{tickers} added. If \code{name} is present, the \code{Watchlist} will be renamed to \code{name} and the \code{tickers} will be added. If \code{watchlist_id} & \code{tickers} are specified and \code{action = NULL} this action is assumed.}
\item{\code{"replace"/"r"}}{ If \code{"replace"} is specified \code{watchlist_id} is \emph{required} and will have it's tickers replaced by those provided to \code{tickers}. If \code{name} is also provided the \code{Watchlist} will be renamed and the \code{tickers} will replace the existing ones.}
\item{\code{"delete"/"d"}}{ If \code{"delete"} is specified with \code{watchlist_id}, the \code{tickers} will be deleted from the \code{Watchlist}. If a \code{name} is also specified, the \code{Watchlist} will also be renamed. If \code{tickers} is \code{NULL} the \code{Watchlist} will be deleted.}
}}

\item{live}{\code{(logical)} TRUE / FALSE if you are connecting to a live account. Default to FALSE, so it will use the paper url if nothing was provided.}
}
\value{
\code{Watchlist} \code{(tibble)} If \code{action = "create"/NULL} a \href{https://alpaca.markets/docs/api-documentation/api-v2/watchlist/#watchlist-entity}{Watchlist} Object or array of \code{Watchlist} objects. For all others \code{action}s, an \href{https://alpaca.markets/docs/api-documentation/api-v2/assets/#asset-entity}{Asset} Object or array of Asset objects with the \code{Watchlist} stored as an attribute retrieved with \code{attr(Asset, "info")}. A \code{Watchlist} object has the following information:
\itemize{
\item{\code{name}}{\code{(character)} The name of the watchlist.}
\item{\code{updated_at}}{\code{(POSIXct)} The timestamp of the last update.}
\item{\code{id}}{\code{(character)} Watchlist ID}
\item{\code{account_id}}{\code{(character)} The account ID associated with the watchlist}
\item{\code{created_at}}{\code{(POSIXct)} The timestamp of the watchlist creation.}
}
}
\description{
Access all \href{https://alpaca.markets/docs/api-documentation/api-v2/watchlist}{Watchlist} endpoints. See link for details.
}
\examples{
# get the existing watchlists
watchlist()
# create a watchlist named test with Microsoft
(wl <- watchlist(name = "test", tickers = "MSFT"))
# See it in the list of watchlists
watchlist()
# Get it
(test <- watchlist("test"))
identical(test,wl)
# Get that Watchlist's info
attr(wl, "info")
# delete that watchlist
watchlist("test", action = "d")
# create it again, this time with Apple and Walmart
watchlist(name = "test", tickers = c("AAPL", "WMT"))
# Update the watchlist name to test2 and replace the symbols with "FB", "AAPL", "AMZN", "NFLX", "GOOG"
(wl <- watchlist("test", name = "test2", tickers = c("FB", "AAPL", "AMZN", "NFLX", "GOOG", "WMT"), action = "r"))
attr(wl, "info")
# Walmart needn't be in there, remove it and rename the watchlist appropriately
(wl <- watchlist("test2", name = "FAANG", ticker = "WMT", action = "d"))
attr(wl, "info")
# A new go-to for gains: "FABANGG"
(wl <- watchlist("FAANG", "FABANGG", ticker = c("BYND", "GOOGL"), action = "a"))
# Back to the original FAANG
(wl <- watchlist("FABANGG", tickers = c("BYND", "GOOGL"), action = "d"))
# rename it appropriately
(wl <- watchlist("FABANGG", "FAANG"))
attr(wl, "info")
# remove it (with partial action argument)
watchlist("FAANG", a = "d")
}
\concept{Watchlist}
