% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{wl_transform}
\alias{wl_transform}
\title{Transform watchlist objects}
\usage{
wl_transform(wl, action, wl_info = NULL)
}
\arguments{
\item{wl}{The watchlist object}
}
\value{
\code{(tibble)} with respective R compliant objects (POSIXct)
}
\description{
Replaces timestamps with POSIXct in watchlist info
}
\keyword{internal}
